Network.AWS.APIGateway.CreateAPIKey

Description

Create an ApiKey resource.

AWS CLI

Creating a Request

createAPIKey :: CreateAPIKey

Creates a value of CreateAPIKey with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cakEnabled - Specifies whether the ApiKey can be used by callers.

cakValue - Specifies a value of the API key.

cakCustomerId - An AWS Marketplace customer identifier , when integrating with the AWS SaaS Marketplace.

cakGenerateDistinctId - Specifies whether (true ) or not (false ) the key identifier is distinct from the created API key value.

cakName - The name of the ApiKey .

cakStageKeys - DEPRECATED FOR USAGE PLANS - Specifies stages associated with the API key.

cakDescription - The description of the ApiKey .

data CreateAPIKey

Request to create an ApiKey resource.

See: createAPIKey smart constructor.

Request Lenses

cakEnabled :: Lens' CreateAPIKey (Maybe Bool)

Specifies whether the ApiKey can be used by callers.

cakValue :: Lens' CreateAPIKey (Maybe Text)

Specifies a value of the API key.

cakCustomerId :: Lens' CreateAPIKey (Maybe Text)

An AWS Marketplace customer identifier , when integrating with the AWS SaaS Marketplace.

cakGenerateDistinctId :: Lens' CreateAPIKey (Maybe Bool)

Specifies whether (true ) or not (false ) the key identifier is distinct from the created API key value.

cakName :: Lens' CreateAPIKey (Maybe Text)

The name of the ApiKey .

cakStageKeys :: Lens' CreateAPIKey [StageKey]

DEPRECATED FOR USAGE PLANS - Specifies stages associated with the API key.

cakDescription :: Lens' CreateAPIKey (Maybe Text)

The description of the ApiKey .

Destructuring the Response

apiKey :: APIKey

Creates a value of APIKey with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

akEnabled - Specifies whether the API Key can be used by callers.

akValue - The value of the API Key.

akCustomerId - An AWS Marketplace customer identifier , when integrating with the AWS SaaS Marketplace.

akCreatedDate - The timestamp when the API Key was created.

akName - The name of the API Key.

akId - The identifier of the API Key.

akStageKeys - A list of Stage resources that are associated with the ApiKey resource.

akLastUpdatedDate - The timestamp when the API Key was last updated.

akDescription - The description of the API Key.

data APIKey

A resource that can be distributed to callers for executing Method resources that require an API key. API keys can be mapped to any Stage on any RestApi , which indicates that the callers with the API key can make requests to that stage.

Use API Keys

See: apiKey smart constructor.

Response Lenses

akEnabled :: Lens' APIKey (Maybe Bool)

Specifies whether the API Key can be used by callers.

akValue :: Lens' APIKey (Maybe Text)

The value of the API Key.

akCustomerId :: Lens' APIKey (Maybe Text)

An AWS Marketplace customer identifier , when integrating with the AWS SaaS Marketplace.

akCreatedDate :: Lens' APIKey (Maybe UTCTime)

The timestamp when the API Key was created.

akName :: Lens' APIKey (Maybe Text)

The name of the API Key.

akId :: Lens' APIKey (Maybe Text)

The identifier of the API Key.

akStageKeys :: Lens' APIKey [Text]

A list of Stage resources that are associated with the ApiKey resource.

akLastUpdatedDate :: Lens' APIKey (Maybe UTCTime)

The timestamp when the API Key was last updated.

akDescription :: Lens' APIKey (Maybe Text)

The description of the API Key.