Network.AWS.APIGateway.CreateDeployment

Description

Creates a Deployment resource, which makes a specified RestApi callable over the internet.

Creating a Request

createDeployment

Arguments

:: Text
cdRestAPIId

-> CreateDeployment

Creates a value of CreateDeployment with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdStageDescription - The description of the Stage resource for the Deployment resource to create.

cdVariables - A map that defines the stage variables for the Stage resource that is associated with the new deployment. Variable names can have alphanumeric and underscore characters, and the values must match [A-Za-z0-9-._~:/?#&=,]+ .

cdCacheClusterSize - Specifies the cache cluster size for the Stage resource specified in the input, if a cache cluster is enabled.

cdCacheClusterEnabled - Enables a cache cluster for the Stage resource specified in the input.

cdStageName - The name of the Stage resource for the Deployment resource to create.

cdDescription - The description for the Deployment resource to create.

cdRestAPIId - The string identifier of the associated RestApi .

data CreateDeployment

Requests Amazon API Gateway to create a Deployment resource.

See: createDeployment smart constructor.

Request Lenses

cdStageDescription :: Lens' CreateDeployment (Maybe Text)

The description of the Stage resource for the Deployment resource to create.

cdVariables :: Lens' CreateDeployment (HashMap Text Text)

A map that defines the stage variables for the Stage resource that is associated with the new deployment. Variable names can have alphanumeric and underscore characters, and the values must match [A-Za-z0-9-._~:/?#&=,]+ .

cdCacheClusterSize :: Lens' CreateDeployment (Maybe CacheClusterSize)

Specifies the cache cluster size for the Stage resource specified in the input, if a cache cluster is enabled.

cdCacheClusterEnabled :: Lens' CreateDeployment (Maybe Bool)

Enables a cache cluster for the Stage resource specified in the input.

cdStageName :: Lens' CreateDeployment (Maybe Text)

The name of the Stage resource for the Deployment resource to create.

cdDescription :: Lens' CreateDeployment (Maybe Text)

The description for the Deployment resource to create.

cdRestAPIId :: Lens' CreateDeployment Text

The string identifier of the associated RestApi .

Destructuring the Response

deployment :: Deployment

Creates a value of Deployment with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dApiSummary - A summary of the RestApi at the date and time that the deployment resource was created.

dCreatedDate - The date and time that the deployment resource was created.

dId - The identifier for the deployment resource.

dDescription - The description for the deployment resource.

data Deployment

An immutable representation of a RestApi resource that can be called by users using Stages . A deployment must be associated with a Stage for it to be callable over the Internet.

To create a deployment, call POST on the Deployments resource of a RestApi . To view, update, or delete a deployment, call GET , PATCH , or DELETE on the specified deployment resource (restapis{restapi_id}deployments{deployment_id} ).RestApi , Deployments , Stage , AWS CLI , AWS SDKs

See: deployment smart constructor.

Response Lenses

dApiSummary :: Lens' Deployment (HashMap Text (HashMap Text MethodSnapshot))

A summary of the RestApi at the date and time that the deployment resource was created.

dCreatedDate :: Lens' Deployment (Maybe UTCTime)

The date and time that the deployment resource was created.

dId :: Lens' Deployment (Maybe Text)

The identifier for the deployment resource.

dDescription :: Lens' Deployment (Maybe Text)

The description for the deployment resource.