Network.AWS.APIGateway.CreateDocumentationVersion

Description

Undocumented operation.

Creating a Request

createDocumentationVersion

Arguments

:: Text
cdvRestAPIId

-> Text
cdvDocumentationVersion

-> CreateDocumentationVersion

Creates a value of CreateDocumentationVersion with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdvStageName - The stage name to be associated with the new documentation snapshot.

cdvDescription - A description about the new documentation snapshot.

cdvRestAPIId - [Required] The string identifier of the associated RestApi .

cdvDocumentationVersion - [Required] The version identifier of the new snapshot.

data CreateDocumentationVersion

Creates a new documentation version of a given API.

See: createDocumentationVersion smart constructor.

Request Lenses

cdvStageName :: Lens' CreateDocumentationVersion (Maybe Text)

The stage name to be associated with the new documentation snapshot.

cdvDescription :: Lens' CreateDocumentationVersion (Maybe Text)

A description about the new documentation snapshot.

cdvRestAPIId :: Lens' CreateDocumentationVersion Text

Required

The string identifier of the associated RestApi .

cdvDocumentationVersion :: Lens' CreateDocumentationVersion Text

Required

The version identifier of the new snapshot.

Destructuring the Response

documentationVersion :: DocumentationVersion

Creates a value of DocumentationVersion with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dvCreatedDate - The date when the API documentation snapshot is created.

dvVersion - The version identifier of the API documentation snapshot.

dvDescription - The description of the API documentation snapshot.

data DocumentationVersion

A snapshot of the documentation of an API.

Publishing API documentation involves creating a documentation version associated with an API stage and exporting the versioned documentation to an external (e.g., Swagger) file.

Documenting an API , DocumentationPart , DocumentationVersions

See: documentationVersion smart constructor.

Response Lenses

dvCreatedDate :: Lens' DocumentationVersion (Maybe UTCTime)

The date when the API documentation snapshot is created.

dvVersion :: Lens' DocumentationVersion (Maybe Text)

The version identifier of the API documentation snapshot.

dvDescription :: Lens' DocumentationVersion (Maybe Text)

The description of the API documentation snapshot.