Network.AWS.APIGateway.CreateDomainName

Description

Creates a new domain name.

Creating a Request

createDomainName

Arguments

:: Text
cdnDomainName

-> CreateDomainName

Creates a value of CreateDomainName with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdnCertificateName - The user-friendly name of the certificate.

cdnCertificateARN - The reference to an AWS-managed certificate. AWS Certificate Manager is the only supported source.

cdnCertificatePrivateKey - [Deprecated] Your certificate's private key.

cdnCertificateBody - [Deprecated] The body of the server certificate provided by your certificate authority.

cdnCertificateChain - [Deprecated] The intermediate certificates and optionally the root certificate, one after the other without any blank lines. If you include the root certificate, your certificate chain must start with intermediate certificates and end with the root certificate. Use the intermediate certificates that were provided by your certificate authority. Do not include any intermediaries that are not in the chain of trust path.

cdnDomainName - (Required) The name of the DomainName resource.

data CreateDomainName

A request to create a new domain name.

See: createDomainName smart constructor.

Request Lenses

cdnCertificateName :: Lens' CreateDomainName (Maybe Text)

The user-friendly name of the certificate.

cdnCertificateARN :: Lens' CreateDomainName (Maybe Text)

The reference to an AWS-managed certificate. AWS Certificate Manager is the only supported source.

cdnCertificatePrivateKey :: Lens' CreateDomainName (Maybe Text)

Deprecated

Your certificate's private key.

cdnCertificateBody :: Lens' CreateDomainName (Maybe Text)

Deprecated

The body of the server certificate provided by your certificate authority.

cdnCertificateChain :: Lens' CreateDomainName (Maybe Text)

Deprecated

The intermediate certificates and optionally the root certificate, one after the other without any blank lines. If you include the root certificate, your certificate chain must start with intermediate certificates and end with the root certificate. Use the intermediate certificates that were provided by your certificate authority. Do not include any intermediaries that are not in the chain of trust path.

cdnDomainName :: Lens' CreateDomainName Text

(Required) The name of the DomainName resource.

Destructuring the Response

domainName :: DomainName

Creates a value of DomainName with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dnCertificateName - The name of the certificate.

dnCertificateARN - The reference to an AWS-managed certificate. AWS Certificate Manager is the only supported source.

dnDomainName - The name of the DomainName resource.

dnCertificateUploadDate - The timestamp when the certificate was uploaded.

dnDistributionDomainName - The domain name of the Amazon CloudFront distribution. For more information, see the Amazon CloudFront documentation .

data DomainName

Represents a domain name that is contained in a simpler, more intuitive URL that can be called.

Use Client-Side Certificate

See: domainName smart constructor.

Response Lenses

dnCertificateName :: Lens' DomainName (Maybe Text)

The name of the certificate.

dnCertificateARN :: Lens' DomainName (Maybe Text)

The reference to an AWS-managed certificate. AWS Certificate Manager is the only supported source.

dnDomainName :: Lens' DomainName (Maybe Text)

The name of the DomainName resource.

dnCertificateUploadDate :: Lens' DomainName (Maybe UTCTime)

The timestamp when the certificate was uploaded.

dnDistributionDomainName :: Lens' DomainName (Maybe Text)

The domain name of the Amazon CloudFront distribution. For more information, see the Amazon CloudFront documentation .