Network.AWS.APIGateway.CreateStage

Description

Creates a new Stage resource that references a pre-existing Deployment for the API.

Creating a Request

createStage

Arguments

:: Text
csRestAPIId

-> Text
csStageName

-> Text
csDeploymentId

-> CreateStage

Creates a value of CreateStage with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

csVariables - A map that defines the stage variables for the new Stage resource. Variable names can have alphanumeric and underscore characters, and the values must match [A-Za-z0-9-._~:/?#&=,]+ .

csDocumentationVersion - The version of the associated API documentation.

csCacheClusterSize - The stage's cache cluster size.

csCacheClusterEnabled - Whether cache clustering is enabled for the stage.

csDescription - The description of the Stage resource.

csRestAPIId - The string identifier of the associated RestApi .

csStageName - The name for the Stage resource.

csDeploymentId - The identifier of the Deployment resource for the Stage resource.

data CreateStage

Requests Amazon API Gateway to create a Stage resource.

See: createStage smart constructor.

Request Lenses

csVariables :: Lens' CreateStage (HashMap Text Text)

A map that defines the stage variables for the new Stage resource. Variable names can have alphanumeric and underscore characters, and the values must match [A-Za-z0-9-._~:/?#&=,]+ .

csDocumentationVersion :: Lens' CreateStage (Maybe Text)

The version of the associated API documentation.

csCacheClusterSize :: Lens' CreateStage (Maybe CacheClusterSize)

The stage's cache cluster size.

csCacheClusterEnabled :: Lens' CreateStage (Maybe Bool)

Whether cache clustering is enabled for the stage.

csDescription :: Lens' CreateStage (Maybe Text)

The description of the Stage resource.

csRestAPIId :: Lens' CreateStage Text

The string identifier of the associated RestApi .

csStageName :: Lens' CreateStage Text

The name for the Stage resource.

csDeploymentId :: Lens' CreateStage Text

The identifier of the Deployment resource for the Stage resource.

Destructuring the Response

stage :: Stage

Creates a value of Stage with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sDeploymentId - The identifier of the Deployment that the stage points to.

sVariables - A map that defines the stage variables for a Stage resource. Variable names can have alphanumeric and underscore characters, and the values must match [A-Za-z0-9-._~:/?#&=,]+ .

sDocumentationVersion - The version of the associated API documentation.

sClientCertificateId - The identifier of a client certificate for an API stage.

sCreatedDate - The timestamp when the stage was created.

sCacheClusterStatus - The status of the cache cluster for the stage, if enabled.

sMethodSettings - A map that defines the method settings for a Stage resource. Keys (designated as /{method_setting_key below) are method paths defined as {resource_path}/{http_method} for an individual method override, or ** for overriding all methods in the stage.

sLastUpdatedDate - The timestamp when the stage last updated.

sCacheClusterSize - The size of the cache cluster for the stage, if enabled.

sCacheClusterEnabled - Specifies whether a cache cluster is enabled for the stage.

sStageName - The name of the stage is the first path segment in the Uniform Resource Identifier (URI) of a call to Amazon API Gateway.

sDescription - The stage's description.

data Stage

Represents a unique identifier for a version of a deployed RestApi that is callable by users.

Deploy an API

See: stage smart constructor.

Response Lenses

sDeploymentId :: Lens' Stage (Maybe Text)

The identifier of the Deployment that the stage points to.

sVariables :: Lens' Stage (HashMap Text Text)

A map that defines the stage variables for a Stage resource. Variable names can have alphanumeric and underscore characters, and the values must match [A-Za-z0-9-._~:/?#&=,]+ .

sDocumentationVersion :: Lens' Stage (Maybe Text)

The version of the associated API documentation.

sClientCertificateId :: Lens' Stage (Maybe Text)

The identifier of a client certificate for an API stage.

sCreatedDate :: Lens' Stage (Maybe UTCTime)

The timestamp when the stage was created.

sCacheClusterStatus :: Lens' Stage (Maybe CacheClusterStatus)

The status of the cache cluster for the stage, if enabled.

sMethodSettings :: Lens' Stage (HashMap Text MethodSetting)

A map that defines the method settings for a Stage resource. Keys (designated as /{method_setting_key below) are method paths defined as {resource_path}/{http_method} for an individual method override, or ** for overriding all methods in the stage.

sLastUpdatedDate :: Lens' Stage (Maybe UTCTime)

The timestamp when the stage last updated.

sCacheClusterSize :: Lens' Stage (Maybe CacheClusterSize)

The size of the cache cluster for the stage, if enabled.

sCacheClusterEnabled :: Lens' Stage (Maybe Bool)

Specifies whether a cache cluster is enabled for the stage.

sStageName :: Lens' Stage (Maybe Text)

The name of the stage is the first path segment in the Uniform Resource Identifier (URI) of a call to Amazon API Gateway.

sDescription :: Lens' Stage (Maybe Text)

The stage's description.