Network.AWS.APIGateway.CreateUsagePlanKey

Description

Creates a usage plan key for adding an existing API key to a usage plan.

Creating a Request

createUsagePlanKey

Arguments

:: Text
cupkUsagePlanId

-> Text
cupkKeyId

-> Text
cupkKeyType

-> CreateUsagePlanKey

Creates a value of CreateUsagePlanKey with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cupkUsagePlanId - The Id of the UsagePlan resource representing the usage plan containing the to-be-created UsagePlanKey resource representing a plan customer.

cupkKeyId - The identifier of a UsagePlanKey resource for a plan customer.

cupkKeyType - The type of a UsagePlanKey resource for a plan customer.

data CreateUsagePlanKey

The POST request to create a usage plan key for adding an existing API key to a usage plan.

See: createUsagePlanKey smart constructor.

Request Lenses

cupkUsagePlanId :: Lens' CreateUsagePlanKey Text

The Id of the UsagePlan resource representing the usage plan containing the to-be-created UsagePlanKey resource representing a plan customer.

cupkKeyId :: Lens' CreateUsagePlanKey Text

The identifier of a UsagePlanKey resource for a plan customer.

cupkKeyType :: Lens' CreateUsagePlanKey Text

The type of a UsagePlanKey resource for a plan customer.

Destructuring the Response

usagePlanKey :: UsagePlanKey

Creates a value of UsagePlanKey with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

upkValue - The value of a usage plan key.

upkName - The name of a usage plan key.

upkId - The Id of a usage plan key.

upkType - The type of a usage plan key. Currently, the valid key type is API_KEY .

data UsagePlanKey

Represents a usage plan key to identify a plan customer.

To associate an API stage with a selected API key in a usage plan, you must create a UsagePlanKey resource to represent the selected ApiKey .

" Create and Use Usage Plans

See: usagePlanKey smart constructor.

Response Lenses

upkValue :: Lens' UsagePlanKey (Maybe Text)

The value of a usage plan key.

upkName :: Lens' UsagePlanKey (Maybe Text)

The name of a usage plan key.

upkId :: Lens' UsagePlanKey (Maybe Text)

The Id of a usage plan key.

upkType :: Lens' UsagePlanKey (Maybe Text)

The type of a usage plan key. Currently, the valid key type is API_KEY .