Network.AWS.APIGateway.DeleteAuthorizer

Description

Deletes an existing Authorizer resource.

AWS CLI

Creating a Request

deleteAuthorizer

Arguments

:: Text
daRestAPIId

-> Text
daAuthorizerId

-> DeleteAuthorizer

Creates a value of DeleteAuthorizer with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

daRestAPIId - The string identifier of the associated RestApi .

daAuthorizerId - The identifier of the Authorizer resource.

data DeleteAuthorizer

Request to delete an existing Authorizer resource.

See: deleteAuthorizer smart constructor.

Request Lenses

daRestAPIId :: Lens' DeleteAuthorizer Text

The string identifier of the associated RestApi .

daAuthorizerId :: Lens' DeleteAuthorizer Text

The identifier of the Authorizer resource.

Destructuring the Response

deleteAuthorizerResponse :: DeleteAuthorizerResponse

Creates a value of DeleteAuthorizerResponse with the minimum fields required to make a request.

data DeleteAuthorizerResponse

See: deleteAuthorizerResponse smart constructor.