Network.AWS.APIGateway.DeleteClientCertificate

Description

Deletes the ClientCertificate resource.

Creating a Request

deleteClientCertificate

Arguments

:: Text
dccClientCertificateId

-> DeleteClientCertificate

Creates a value of DeleteClientCertificate with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dccClientCertificateId - The identifier of the ClientCertificate resource to be deleted.

data DeleteClientCertificate

A request to delete the ClientCertificate resource.

See: deleteClientCertificate smart constructor.

Request Lenses

dccClientCertificateId :: Lens' DeleteClientCertificate Text

The identifier of the ClientCertificate resource to be deleted.

Destructuring the Response

deleteClientCertificateResponse :: DeleteClientCertificateResponse

Creates a value of DeleteClientCertificateResponse with the minimum fields required to make a request.

data DeleteClientCertificateResponse

See: deleteClientCertificateResponse smart constructor.