Network.AWS.APIGateway.DeleteMethodResponse

Description

Deletes an existing MethodResponse resource.

Creating a Request

deleteMethodResponse

Arguments

:: Text
dmRestAPIId

-> Text
dmResourceId

-> Text
dmHttpMethod

-> Text
dmStatusCode

-> DeleteMethodResponse

Creates a value of DeleteMethodResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dmRestAPIId - The string identifier of the associated RestApi .

dmResourceId - The Resource identifier for the MethodResponse resource.

dmHttpMethod - The HTTP verb of the Method resource.

dmStatusCode - The status code identifier for the MethodResponse resource.

data DeleteMethodResponse

A request to delete an existing MethodResponse resource.

See: deleteMethodResponse smart constructor.

Request Lenses

dmRestAPIId :: Lens' DeleteMethodResponse Text

The string identifier of the associated RestApi .

dmResourceId :: Lens' DeleteMethodResponse Text

The Resource identifier for the MethodResponse resource.

dmHttpMethod :: Lens' DeleteMethodResponse Text

The HTTP verb of the Method resource.

dmStatusCode :: Lens' DeleteMethodResponse Text

The status code identifier for the MethodResponse resource.

Destructuring the Response

deleteMethodResponseResponse :: DeleteMethodResponseResponse

Creates a value of DeleteMethodResponseResponse with the minimum fields required to make a request.

data DeleteMethodResponseResponse

See: deleteMethodResponseResponse smart constructor.