Network.AWS.APIGateway.DeleteRestAPI

Description

Deletes the specified API.

Creating a Request

deleteRestAPI

Arguments

:: Text
draRestAPIId

-> DeleteRestAPI

Creates a value of DeleteRestAPI with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

draRestAPIId - The string identifier of the associated RestApi .

data DeleteRestAPI

Request to delete the specified API from your collection.

See: deleteRestAPI smart constructor.

Request Lenses

draRestAPIId :: Lens' DeleteRestAPI Text

The string identifier of the associated RestApi .

Destructuring the Response

deleteRestAPIResponse :: DeleteRestAPIResponse

Creates a value of DeleteRestAPIResponse with the minimum fields required to make a request.

data DeleteRestAPIResponse

See: deleteRestAPIResponse smart constructor.