Network.AWS.APIGateway.DeleteStage

Description

Deletes a Stage resource.

Creating a Request

deleteStage

Arguments

:: Text
dsRestAPIId

-> Text
dsStageName

-> DeleteStage

Creates a value of DeleteStage with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsRestAPIId - The string identifier of the associated RestApi .

dsStageName - The name of the Stage resource to delete.

data DeleteStage

Requests Amazon API Gateway to delete a Stage resource.

See: deleteStage smart constructor.

Request Lenses

dsRestAPIId :: Lens' DeleteStage Text

The string identifier of the associated RestApi .

dsStageName :: Lens' DeleteStage Text

The name of the Stage resource to delete.

Destructuring the Response

deleteStageResponse :: DeleteStageResponse

Creates a value of DeleteStageResponse with the minimum fields required to make a request.

data DeleteStageResponse

See: deleteStageResponse smart constructor.