Network.AWS.APIGateway.DeleteUsagePlanKey

Description

Deletes a usage plan key and remove the underlying API key from the associated usage plan.

Creating a Request

deleteUsagePlanKey

Arguments

:: Text
dupkUsagePlanId

-> Text
dupkKeyId

-> DeleteUsagePlanKey

Creates a value of DeleteUsagePlanKey with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dupkUsagePlanId - The Id of the UsagePlan resource representing the usage plan containing the to-be-deleted UsagePlanKey resource representing a plan customer.

dupkKeyId - The Id of the UsagePlanKey resource to be deleted.

data DeleteUsagePlanKey

The DELETE request to delete a usage plan key and remove the underlying API key from the associated usage plan.

See: deleteUsagePlanKey smart constructor.

Request Lenses

dupkUsagePlanId :: Lens' DeleteUsagePlanKey Text

The Id of the UsagePlan resource representing the usage plan containing the to-be-deleted UsagePlanKey resource representing a plan customer.

dupkKeyId :: Lens' DeleteUsagePlanKey Text

The Id of the UsagePlanKey resource to be deleted.

Destructuring the Response

deleteUsagePlanKeyResponse :: DeleteUsagePlanKeyResponse

Creates a value of DeleteUsagePlanKeyResponse with the minimum fields required to make a request.

data DeleteUsagePlanKeyResponse

See: deleteUsagePlanKeyResponse smart constructor.