Network.AWS.APIGateway.GetAPIKey

Description

Gets information about the current ApiKey resource.

Creating a Request

getAPIKey

Arguments

:: Text
gakApiKey

-> GetAPIKey

Creates a value of GetAPIKey with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gakIncludeValue - A boolean flag to specify whether (true ) or not (false ) the result contains the key value.

gakApiKey - The identifier of the ApiKey resource.

data GetAPIKey

A request to get information about the current ApiKey resource.

See: getAPIKey smart constructor.

Request Lenses

gakIncludeValue :: Lens' GetAPIKey (Maybe Bool)

A boolean flag to specify whether (true ) or not (false ) the result contains the key value.

gakApiKey :: Lens' GetAPIKey Text

The identifier of the ApiKey resource.

Destructuring the Response

apiKey :: APIKey

Creates a value of APIKey with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

akEnabled - Specifies whether the API Key can be used by callers.

akValue - The value of the API Key.

akCustomerId - An AWS Marketplace customer identifier , when integrating with the AWS SaaS Marketplace.

akCreatedDate - The timestamp when the API Key was created.

akName - The name of the API Key.

akId - The identifier of the API Key.

akStageKeys - A list of Stage resources that are associated with the ApiKey resource.

akLastUpdatedDate - The timestamp when the API Key was last updated.

akDescription - The description of the API Key.

data APIKey

A resource that can be distributed to callers for executing Method resources that require an API key. API keys can be mapped to any Stage on any RestApi , which indicates that the callers with the API key can make requests to that stage.

Use API Keys

See: apiKey smart constructor.

Response Lenses

akEnabled :: Lens' APIKey (Maybe Bool)

Specifies whether the API Key can be used by callers.

akValue :: Lens' APIKey (Maybe Text)

The value of the API Key.

akCustomerId :: Lens' APIKey (Maybe Text)

An AWS Marketplace customer identifier , when integrating with the AWS SaaS Marketplace.

akCreatedDate :: Lens' APIKey (Maybe UTCTime)

The timestamp when the API Key was created.

akName :: Lens' APIKey (Maybe Text)

The name of the API Key.

akId :: Lens' APIKey (Maybe Text)

The identifier of the API Key.

akStageKeys :: Lens' APIKey [Text]

A list of Stage resources that are associated with the ApiKey resource.

akLastUpdatedDate :: Lens' APIKey (Maybe UTCTime)

The timestamp when the API Key was last updated.

akDescription :: Lens' APIKey (Maybe Text)

The description of the API Key.