Network.AWS.APIGateway.GetAPIKeys

Description

Gets information about the current ApiKeys resource.

This operation returns paginated results.

Creating a Request

getAPIKeys :: GetAPIKeys

Creates a value of GetAPIKeys with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gakIncludeValues - A boolean flag to specify whether (true ) or not (false ) the result contains key values.

gakCustomerId - The identifier of a customer in AWS Marketplace or an external system, such as a developer portal.

gakNameQuery - The name of queried API keys.

gakLimit - The maximum number of returned results per page.

gakPosition - The current pagination position in the paged result set.

data GetAPIKeys

A request to get information about the current ApiKeys resource.

See: getAPIKeys smart constructor.

Request Lenses

gakIncludeValues :: Lens' GetAPIKeys (Maybe Bool)

A boolean flag to specify whether (true ) or not (false ) the result contains key values.

gakCustomerId :: Lens' GetAPIKeys (Maybe Text)

The identifier of a customer in AWS Marketplace or an external system, such as a developer portal.

gakNameQuery :: Lens' GetAPIKeys (Maybe Text)

The name of queried API keys.

gakLimit :: Lens' GetAPIKeys (Maybe Int)

The maximum number of returned results per page.

gakPosition :: Lens' GetAPIKeys (Maybe Text)

The current pagination position in the paged result set.

Destructuring the Response

getAPIKeysResponse

Arguments

:: Int
gakrsResponseStatus

-> GetAPIKeysResponse

Creates a value of GetAPIKeysResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gakrsItems - The current page of elements from this collection.

gakrsWarnings - A list of warning messages logged during the import of API keys when the failOnWarnings option is set to true.

gakrsPosition - Undocumented member.

gakrsResponseStatus - -- | The response status code.

data GetAPIKeysResponse

Represents a collection of API keys as represented by an ApiKeys resource.

Use API Keys

See: getAPIKeysResponse smart constructor.

Response Lenses

gakrsItems :: Lens' GetAPIKeysResponse [APIKey]

The current page of elements from this collection.

gakrsWarnings :: Lens' GetAPIKeysResponse [Text]

A list of warning messages logged during the import of API keys when the failOnWarnings option is set to true.

gakrsPosition :: Lens' GetAPIKeysResponse (Maybe Text)

Undocumented member.

gakrsResponseStatus :: Lens' GetAPIKeysResponse Int

- | The response status code.