Network.AWS.APIGateway.GetAuthorizers

Description

Describe an existing Authorizers resource.

AWS CLI

Creating a Request

getAuthorizers

Arguments

:: Text
gaRestAPIId

-> GetAuthorizers

Creates a value of GetAuthorizers with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gaLimit - The maximum number of returned results per page.

gaPosition - The current pagination position in the paged result set.

gaRestAPIId - The string identifier of the associated RestApi .

data GetAuthorizers

Request to describe an existing Authorizers resource.

See: getAuthorizers smart constructor.

Request Lenses

gaLimit :: Lens' GetAuthorizers (Maybe Int)

The maximum number of returned results per page.

gaPosition :: Lens' GetAuthorizers (Maybe Text)

The current pagination position in the paged result set.

gaRestAPIId :: Lens' GetAuthorizers Text

The string identifier of the associated RestApi .

Destructuring the Response

getAuthorizersResponse

Arguments

:: Int
garsResponseStatus

-> GetAuthorizersResponse

Creates a value of GetAuthorizersResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

garsItems - The current page of elements from this collection.

garsPosition - Undocumented member.

garsResponseStatus - -- | The response status code.

data GetAuthorizersResponse

Represents a collection of Authorizer resources.

Enable custom authorization

See: getAuthorizersResponse smart constructor.

Response Lenses

garsItems :: Lens' GetAuthorizersResponse [Authorizer]

The current page of elements from this collection.

garsPosition :: Lens' GetAuthorizersResponse (Maybe Text)

Undocumented member.

garsResponseStatus :: Lens' GetAuthorizersResponse Int

- | The response status code.