Network.AWS.APIGateway.GetClientCertificate

Description

Gets information about the current ClientCertificate resource.

Creating a Request

getClientCertificate

Arguments

:: Text
gccClientCertificateId

-> GetClientCertificate

Creates a value of GetClientCertificate with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gccClientCertificateId - The identifier of the ClientCertificate resource to be described.

data GetClientCertificate

A request to get information about the current ClientCertificate resource.

See: getClientCertificate smart constructor.

Request Lenses

gccClientCertificateId :: Lens' GetClientCertificate Text

The identifier of the ClientCertificate resource to be described.

Destructuring the Response

clientCertificate :: ClientCertificate

Creates a value of ClientCertificate with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ccPemEncodedCertificate - The PEM-encoded public key of the client certificate, which can be used to configure certificate authentication in the integration endpoint .

ccClientCertificateId - The identifier of the client certificate.

ccCreatedDate - The timestamp when the client certificate was created.

ccExpirationDate - The timestamp when the client certificate will expire.

ccDescription - The description of the client certificate.

data ClientCertificate

Represents a client certificate used to configure client-side SSL authentication while sending requests to the integration endpoint.

Client certificates are used authenticate an API by the back-end server. To authenticate an API client (or user), use a custom Authorizer .Use Client-Side Certificate

See: clientCertificate smart constructor.

Response Lenses

ccPemEncodedCertificate :: Lens' ClientCertificate (Maybe Text)

The PEM-encoded public key of the client certificate, which can be used to configure certificate authentication in the integration endpoint .

ccClientCertificateId :: Lens' ClientCertificate (Maybe Text)

The identifier of the client certificate.

ccCreatedDate :: Lens' ClientCertificate (Maybe UTCTime)

The timestamp when the client certificate was created.

ccExpirationDate :: Lens' ClientCertificate (Maybe UTCTime)

The timestamp when the client certificate will expire.

ccDescription :: Lens' ClientCertificate (Maybe Text)

The description of the client certificate.