Network.AWS.APIGateway.GetClientCertificates

Description

Gets a collection of ClientCertificate resources.

This operation returns paginated results.

Creating a Request

getClientCertificates :: GetClientCertificates

Creates a value of GetClientCertificates with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gccLimit - The maximum number of returned results per page. The value is 25 by default and could be between 1 - 500.

gccPosition - The current pagination position in the paged result set.

data GetClientCertificates

A request to get information about a collection of ClientCertificate resources.

See: getClientCertificates smart constructor.

Request Lenses

gccLimit :: Lens' GetClientCertificates (Maybe Int)

The maximum number of returned results per page. The value is 25 by default and could be between 1 - 500.

gccPosition :: Lens' GetClientCertificates (Maybe Text)

The current pagination position in the paged result set.

Destructuring the Response

getClientCertificatesResponse

Arguments

:: Int
gccrsResponseStatus

-> GetClientCertificatesResponse

Creates a value of GetClientCertificatesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gccrsItems - The current page of elements from this collection.

gccrsPosition - Undocumented member.

gccrsResponseStatus - -- | The response status code.

data GetClientCertificatesResponse

Represents a collection of ClientCertificate resources.

Use Client-Side Certificate

See: getClientCertificatesResponse smart constructor.

Response Lenses

gccrsItems :: Lens' GetClientCertificatesResponse [ClientCertificate]

The current page of elements from this collection.

gccrsPosition :: Lens' GetClientCertificatesResponse (Maybe Text)

Undocumented member.

gccrsResponseStatus :: Lens' GetClientCertificatesResponse Int

- | The response status code.