Network.AWS.APIGateway.GetDeployments

Description

Gets information about a Deployments collection.

This operation returns paginated results.

Creating a Request

getDeployments

Arguments

:: Text
gdRestAPIId

-> GetDeployments

Creates a value of GetDeployments with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gdLimit - The maximum number of returned results per page. The value is 25 by default and could be between 1 - 500.

gdPosition - The current pagination position in the paged result set.

gdRestAPIId - The string identifier of the associated RestApi .

data GetDeployments

Requests Amazon API Gateway to get information about a Deployments collection.

See: getDeployments smart constructor.

Request Lenses

gdLimit :: Lens' GetDeployments (Maybe Int)

The maximum number of returned results per page. The value is 25 by default and could be between 1 - 500.

gdPosition :: Lens' GetDeployments (Maybe Text)

The current pagination position in the paged result set.

gdRestAPIId :: Lens' GetDeployments Text

The string identifier of the associated RestApi .

Destructuring the Response

getDeploymentsResponse

Arguments

:: Int
gdrsResponseStatus

-> GetDeploymentsResponse

Creates a value of GetDeploymentsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gdrsItems - The current page of elements from this collection.

gdrsPosition - Undocumented member.

gdrsResponseStatus - -- | The response status code.

data GetDeploymentsResponse

Represents a collection resource that contains zero or more references to your existing deployments, and links that guide you on how to interact with your collection. The collection offers a paginated view of the contained deployments.

To create a new deployment of a RestApi , make a POST request against this resource. To view, update, or delete an existing deployment, make a GET , PATCH , or DELETE request, respectively, on a specified Deployment resource.Deploying an API , AWS CLI , AWS SDKs

See: getDeploymentsResponse smart constructor.

Response Lenses

gdrsItems :: Lens' GetDeploymentsResponse [Deployment]

The current page of elements from this collection.

gdrsPosition :: Lens' GetDeploymentsResponse (Maybe Text)

Undocumented member.

gdrsResponseStatus :: Lens' GetDeploymentsResponse Int

- | The response status code.