Network.AWS.APIGateway.GetDocumentationParts

Description

Undocumented operation.

Creating a Request

getDocumentationParts

Arguments

:: Text
gdpRestAPIId

-> GetDocumentationParts

Creates a value of GetDocumentationParts with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gdpPath - The path of API entities of the to-be-retrieved documentation parts.

gdpNameQuery - The name of API entities of the to-be-retrieved documentation parts.

gdpLimit - The maximum number of returned results per page.

gdpType - The type of API entities of the to-be-retrieved documentation parts.

gdpPosition - The current pagination position in the paged result set.

gdpRestAPIId - [Required] The string identifier of the associated RestApi .

data GetDocumentationParts

Gets the documentation parts of an API. The result may be filtered by the type, name, or path of API entities (targets).

See: getDocumentationParts smart constructor.

Request Lenses

gdpPath :: Lens' GetDocumentationParts (Maybe Text)

The path of API entities of the to-be-retrieved documentation parts.

gdpNameQuery :: Lens' GetDocumentationParts (Maybe Text)

The name of API entities of the to-be-retrieved documentation parts.

gdpLimit :: Lens' GetDocumentationParts (Maybe Int)

The maximum number of returned results per page.

gdpType :: Lens' GetDocumentationParts (Maybe DocumentationPartType)

The type of API entities of the to-be-retrieved documentation parts.

gdpPosition :: Lens' GetDocumentationParts (Maybe Text)

The current pagination position in the paged result set.

gdpRestAPIId :: Lens' GetDocumentationParts Text

Required

The string identifier of the associated RestApi .

Destructuring the Response

getDocumentationPartsResponse

Arguments

:: Int
gdprsResponseStatus

-> GetDocumentationPartsResponse

Creates a value of GetDocumentationPartsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gdprsItems - The current page of elements from this collection.

gdprsPosition - Undocumented member.

gdprsResponseStatus - -- | The response status code.

data GetDocumentationPartsResponse

The collection of documentation parts of an API.

Documenting an API , DocumentationPart

See: getDocumentationPartsResponse smart constructor.

Response Lenses

gdprsItems :: Lens' GetDocumentationPartsResponse [DocumentationPart]

The current page of elements from this collection.

gdprsPosition :: Lens' GetDocumentationPartsResponse (Maybe Text)

Undocumented member.

gdprsResponseStatus :: Lens' GetDocumentationPartsResponse Int

- | The response status code.