Network.AWS.APIGateway.GetDocumentationVersion

Description

Undocumented operation.

Creating a Request

getDocumentationVersion

Arguments

:: Text
gdvdRestAPIId

-> Text
gdvdDocumentationVersion

-> GetDocumentationVersion

Creates a value of GetDocumentationVersion with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gdvdRestAPIId - [Required] The string identifier of the associated RestApi .

gdvdDocumentationVersion - [Required] The version identifier of the to-be-retrieved documentation snapshot.

data GetDocumentationVersion

Gets a documentation snapshot of an API.

See: getDocumentationVersion smart constructor.

Request Lenses

gdvdRestAPIId :: Lens' GetDocumentationVersion Text

Required

The string identifier of the associated RestApi .

gdvdDocumentationVersion :: Lens' GetDocumentationVersion Text

Required

The version identifier of the to-be-retrieved documentation snapshot.

Destructuring the Response

documentationVersion :: DocumentationVersion

Creates a value of DocumentationVersion with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dvCreatedDate - The date when the API documentation snapshot is created.

dvVersion - The version identifier of the API documentation snapshot.

dvDescription - The description of the API documentation snapshot.

data DocumentationVersion

A snapshot of the documentation of an API.

Publishing API documentation involves creating a documentation version associated with an API stage and exporting the versioned documentation to an external (e.g., Swagger) file.

Documenting an API , DocumentationPart , DocumentationVersions

See: documentationVersion smart constructor.

Response Lenses

dvCreatedDate :: Lens' DocumentationVersion (Maybe UTCTime)

The date when the API documentation snapshot is created.

dvVersion :: Lens' DocumentationVersion (Maybe Text)

The version identifier of the API documentation snapshot.

dvDescription :: Lens' DocumentationVersion (Maybe Text)

The description of the API documentation snapshot.