Network.AWS.APIGateway.GetDocumentationVersions

Description

Undocumented operation.

Creating a Request

getDocumentationVersions

Arguments

:: Text
gdvRestAPIId

-> GetDocumentationVersions

Creates a value of GetDocumentationVersions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gdvLimit - The maximum number of returned results per page.

gdvPosition - The current pagination position in the paged result set.

gdvRestAPIId - [Required] The string identifier of the associated RestApi .

data GetDocumentationVersions

Gets the documentation versions of an API.

See: getDocumentationVersions smart constructor.

Request Lenses

gdvLimit :: Lens' GetDocumentationVersions (Maybe Int)

The maximum number of returned results per page.

gdvPosition :: Lens' GetDocumentationVersions (Maybe Text)

The current pagination position in the paged result set.

gdvRestAPIId :: Lens' GetDocumentationVersions Text

Required

The string identifier of the associated RestApi .

Destructuring the Response

getDocumentationVersionsResponse

Arguments

:: Int
gdvrsResponseStatus

-> GetDocumentationVersionsResponse

Creates a value of GetDocumentationVersionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gdvrsItems - The current page of elements from this collection.

gdvrsPosition - Undocumented member.

gdvrsResponseStatus - -- | The response status code.

data GetDocumentationVersionsResponse

The collection of documentation snapshots of an API.

Use the DocumentationVersions to manage documentation snapshots associated with various API stages.

Documenting an API , DocumentationPart , DocumentationVersion

See: getDocumentationVersionsResponse smart constructor.

Response Lenses

gdvrsItems :: Lens' GetDocumentationVersionsResponse [DocumentationVersion]

The current page of elements from this collection.

gdvrsPosition :: Lens' GetDocumentationVersionsResponse (Maybe Text)

Undocumented member.

gdvrsResponseStatus :: Lens' GetDocumentationVersionsResponse Int

- | The response status code.