Network.AWS.APIGateway.GetExport

Description

Exports a deployed version of a RestApi in a specified format.

Creating a Request

getExport

Arguments

:: Text
geRestAPIId

-> Text
geStageName

-> Text
geExportType

-> GetExport

Creates a value of GetExport with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

geParameters - A key-value map of query string parameters that specify properties of the export, depending on the requested exportType . For exportType swagger , any combination of the following parameters are supported: integrations will export the API with x-amazon-apigateway-integration extensions. authorizers will export the API with x-amazon-apigateway-authorizer extensions. postman will export the API with Postman extensions, allowing for import to the Postman tool

geAccepts - The content-type of the export, for example application/json . Currently application/json and application/yaml are supported for exportType of swagger . This should be specified in the Accept header for direct API requests.

geRestAPIId - The string identifier of the associated RestApi .

geStageName - The name of the Stage that will be exported.

geExportType - The type of export. Currently only swagger is supported.

data GetExport

Request a new export of a RestApi for a particular Stage .

See: getExport smart constructor.

Request Lenses

geParameters :: Lens' GetExport (HashMap Text Text)

A key-value map of query string parameters that specify properties of the export, depending on the requested exportType . For exportType swagger , any combination of the following parameters are supported: integrations will export the API with x-amazon-apigateway-integration extensions. authorizers will export the API with x-amazon-apigateway-authorizer extensions. postman will export the API with Postman extensions, allowing for import to the Postman tool

geAccepts :: Lens' GetExport (Maybe Text)

The content-type of the export, for example application/json . Currently application/json and application/yaml are supported for exportType of swagger . This should be specified in the Accept header for direct API requests.

geRestAPIId :: Lens' GetExport Text

The string identifier of the associated RestApi .

geStageName :: Lens' GetExport Text

The name of the Stage that will be exported.

geExportType :: Lens' GetExport Text

The type of export. Currently only swagger is supported.

Destructuring the Response

getExportResponse

Arguments

:: Int
gersResponseStatus

-> GetExportResponse

Creates a value of GetExportResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gersBody - The binary blob response to GetExport , which contains the export.

gersContentDisposition - The content-disposition header value in the HTTP response.

gersContentType - The content-type header value in the HTTP response. This will correspond to a valid accept type in the request.

gersResponseStatus - -- | The response status code.

data GetExportResponse

The binary blob response to GetExport , which contains the generated SDK.

See: getExportResponse smart constructor.

Response Lenses

gersBody :: Lens' GetExportResponse (Maybe (HashMap Text Value))

The binary blob response to GetExport , which contains the export.

gersContentDisposition :: Lens' GetExportResponse (Maybe Text)

The content-disposition header value in the HTTP response.

gersContentType :: Lens' GetExportResponse (Maybe Text)

The content-type header value in the HTTP response. This will correspond to a valid accept type in the request.

gersResponseStatus :: Lens' GetExportResponse Int

- | The response status code.