Network.AWS.APIGateway.GetMethod

Description

Describe an existing Method resource.

Creating a Request

getMethod

Arguments

:: Text
gmmRestAPIId

-> Text
gmmResourceId

-> Text
gmmHttpMethod

-> GetMethod

Creates a value of GetMethod with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gmmRestAPIId - The string identifier of the associated RestApi .

gmmResourceId - The Resource identifier for the Method resource.

gmmHttpMethod - Specifies the method request's HTTP method type.

data GetMethod

Request to describe an existing Method resource.

See: getMethod smart constructor.

Request Lenses

gmmRestAPIId :: Lens' GetMethod Text

The string identifier of the associated RestApi .

gmmResourceId :: Lens' GetMethod Text

The Resource identifier for the Method resource.

gmmHttpMethod :: Lens' GetMethod Text

Specifies the method request's HTTP method type.

Destructuring the Response

method :: Method

Creates a value of Method with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mMethodResponses - Gets a method response associated with a given HTTP status code. The collection of method responses are encapsulated in a key-value map, where the key is a response's HTTP status code and the value is a MethodResponse resource that specifies the response returned to the caller from the back end through the integration response. Example: Get a 200 OK response of a GET method Request @GET restapisuojnr9hd57resources0cjtchmethodsGETresponses200 HTTP1.1 Content-Type: applicationjson Host: apigateway.us-east-1.amazonaws.com Content-Length: 117 X-Amz-Date: 20160613T215008Z Authorization: AWS4-HMAC-SHA256 Credential={access_key_ID}20160613us-east-1apigatewayaws4_request, SignedHeaders=content-type;host;x-amz-date, Signature={sig4_hash} Response The successful response returns a 200 OK status code and a payload similar to the following: { "_links": { "curies": { "href": "http://docs.aws.amazon.com/apigateway/latest/developerguide/restapi-method-response-{rel}.html", "name": "methodresponse", "templated": true }, "self": { "href": "restapisuojnr9hd57resources0cjtchmethodsGETresponses200", "title": "200" }, "methodresponse:delete": { "href": "restapisuojnr9hd57resources0cjtchmethodsGETresponses200" }, "methodresponse:update": { "href": "restapisuojnr9hd57resources0cjtchmethodsGETresponses200" } }, "responseModels": { "application/json": Empty }, "responseParameters": { "method.response.header.operator": false, "method.response.header.operand_2": false, "method.response.header.operand_1": false }, "statusCode": "200" } @ AWS CLI

mHttpMethod - The method's HTTP verb.

mRequestValidatorId - The identifier of a RequestValidator for request validation.

mRequestModels - A key-value map specifying data schemas, represented by Model resources, (as the mapped value) of the request payloads of given content types (as the mapping key).

mRequestParameters - A key-value map defining required or optional method request parameters that can be accepted by Amazon API Gateway. A key is a method request parameter name matching the pattern of method.request.{location}.{name} , where location is querystring , path , or header and name is a valid and unique parameter name. The value associated with the key is a Boolean flag indicating whether the parameter is required (true ) or optional (false ). The method request parameter names defined here are available in Integration to be mapped to integration request parameters or templates.

mAuthorizerId - The identifier of an Authorizer to use on this method. The authorizationType must be CUSTOM .

mOperationName - A human-friendly operation identifier for the method. For example, you can assign the operationName of ListPets for the GET /pets method in PetStore example.

mAuthorizationType - The method's authorization type. Valid values are NONE for open access, AWS_IAM for using AWS IAM permissions, CUSTOM for using a custom authorizer, or COGNITO_USER_POOLS for using a Cognito user pool.

mApiKeyRequired - A boolean flag specifying whether a valid ApiKey is required to invoke this method.

mMethodIntegration - Gets the method's integration responsible for passing the client-submitted request to the back end and performing necessary transformations to make the request compliant with the back end. Example: Request @GET restapisuojnr9hd57resources0cjtchmethodsGETintegration HTTP1.1 Content-Type: applicationjson Host: apigateway.us-east-1.amazonaws.com Content-Length: 117 X-Amz-Date: 20160613T213210Z Authorization: AWS4-HMAC-SHA256 Credential={access_key_ID}20160613us-east-1apigateway/aws4_request, SignedHeaders=content-type;host;x-amz-date, Signature={sig4_hash} Response The successful response returns a 200 OK status code and a payload similar to the following: { "_links": { "curies": [ { "href": "http://docs.aws.amazon.com/apigateway/latest/developerguide/restapi-integration-{rel}.html", "name": "integration", "templated": true }, { "href": "http://docs.aws.amazon.com/apigateway/latest/developerguide/restapi-integration-response-{rel}.html", "name": "integrationresponse", "templated": true } ], "self": { "href": "restapisuojnr9hd57resources0cjtchmethodsGETintegration" }, "integration:delete": { "href": "restapisuojnr9hd57resources0cjtchmethodsGETintegration" }, "integration:responses": { "href": "restapisuojnr9hd57resources0cjtchmethodsGETintegrationresponses200", "name": "200", "title": "200" }, "integration:update": { "href": "restapisuojnr9hd57resources0cjtchmethodsGETintegration" }, "integrationresponse:put": { "href": "restapisuojnr9hd57resources0cjtchmethodsGETintegrationresponses{status_code}", "templated": true } }, "cacheKeyParameters": [], "cacheNamespace": "0cjtch", "credentials": "arn:aws:iam::123456789012:roleapigAwsProxyRole", "httpMethod": POST, "passthroughBehavior": WHEN_NO_MATCH, "requestTemplates": { "applicationjson": "{n "a": "$input.params(operand1)",n "b": "$input.params(operand2)", n "op": "$input.params(operator)" n}" }, "type": AWS, "uri": "arn:aws:apigateway:us-west-2:lambda:path2015-03-31functionsarn:aws:lambda:us-west-2:123456789012:function:Calcinvocations", "_embedded": { "integration:responses": { "_links": { "self": { "href": "restapisuojnr9hd57resources0cjtchmethodsGETintegrationresponses200", "name": "200", "title": "200" }, "integrationresponse:delete": { "href": "restapisuojnr9hd57resources0cjtchmethodsGETintegrationresponses200" }, "integrationresponse:update": { "href": "restapisuojnr9hd57resources0cjtchmethodsGETintegrationresponses200" } }, "responseParameters": { "method.response.header.operator": "integration.response.body.op", "method.response.header.operand_2": "integration.response.body.b", "method.response.header.operand_1": "integration.response.body.a" }, "responseTemplates": { "applicationjson": "#set($res = $input.path($))n{n "result": "$res.a, $res.b, $res.op => $res.c",n "a" : "$res.a",n "b" : "$res.b",n "op" : "$res.op",n "c" : "$res.c"n}" }, "selectionPattern": "", "statusCode": "200" } } } @ AWS CLI

data Method

Represents a client-facing interface by which the client calls the API to access back-end resources. A Method resource is integrated with an Integration resource. Both consist of a request and one or more responses. The method request takes the client input that is passed to the back end through the integration request. A method response returns the output from the back end to the client through an integration response. A method request is embodied in a Method resource, whereas an integration request is embodied in an Integration resource. On the other hand, a method response is represented by a MethodResponse resource, whereas an integration response is represented by an IntegrationResponse resource.

Example: Retrive the GET method on a specified resource Request The following example request retrieves the information about the GET method on an API resource (3kzxbg5sa2 ) of an API (fugvjdxtri ).

@GET restapisfugvjdxtriresources3kzxbg5sa2methodsGET HTTP1.1 Content-Type: applicationjson Host: apigateway.us-east-1.amazonaws.com X-Amz-Date: 20160603T210259Z Authorization: AWS4-HMAC-SHA256 Credential={access_key_ID}20160603us-east-1apigatewayaws4_request, SignedHeaders=content-type;host;x-amz-date, Signature={sig4_hash} Response The successful response returns a 200 OK@ status code and a payload similar to the following:

@{ "_links": { "curies": [ { "href": "http://docs.aws.amazon.com/apigateway/latest/developerguide/restapi-integration-{rel}.html", "name": "integration", "templated": true }, { "href": "http://docs.aws.amazon.com/apigateway/latest/developerguide/restapi-integration-response-{rel}.html", "name": "integrationresponse", "templated": true }, { "href": "http://docs.aws.amazon.com/apigateway/latest/developerguide/restapi-method-{rel}.html", "name": "method", "templated": true }, { "href": "http://docs.aws.amazon.com/apigateway/latest/developerguide/restapi-method-response-{rel}.html", "name": "methodresponse", "templated": true } ], "self": { "href": "restapisfugvjdxtriresources3kzxbg5sa2methodsGET", "name": GET, "title": GET }, "integration:put": { "href": "restapisfugvjdxtriresources3kzxbg5sa2methodsGETintegration" }, "method:delete": { "href": "restapisfugvjdxtriresources3kzxbg5sa2methodsGET" }, "method:integration": { "href": "restapisfugvjdxtriresources3kzxbg5sa2methodsGETintegration" }, "method:responses": { "href": "restapisfugvjdxtriresources3kzxbg5sa2methodsGETresponses200", "name": "200", "title": "200" }, "method:update": { "href": "restapisfugvjdxtriresources3kzxbg5sa2methodsGET" }, "methodresponse:put": { "href": "restapisfugvjdxtriresources3kzxbg5sa2methodsGETresponses{status_code}", "templated": true } }, "apiKeyRequired": true, "authorizationType": NONE, "httpMethod": GET, "_embedded": { "method:integration": { "_links": { "self": { "href": "restapisfugvjdxtriresources3kzxbg5sa2methodsGETintegration" }, "integration:delete": { "href": "restapisfugvjdxtriresources3kzxbg5sa2methodsGETintegration" }, "integration:responses": { "href": "restapisfugvjdxtriresources3kzxbg5sa2methodsGETintegrationresponses200", "name": "200", "title": "200" }, "integration:update": { "href": "restapisfugvjdxtriresources3kzxbg5sa2methodsGETintegration" }, "integrationresponse:put": { "href": "restapisfugvjdxtriresources3kzxbg5sa2methodsGETintegrationresponses{status_code}", "templated": true } }, "cacheKeyParameters": [], "cacheNamespace": "3kzxbg5sa2", "credentials": "arn:aws:iam::123456789012:roleapigAwsProxyRole", "httpMethod": POST, "passthroughBehavior": WHEN_NO_MATCH, "requestParameters": { "integration.request.header.Content-Type": "'application/x-amz-json-1.1'" }, "requestTemplates": { "applicationjson": "{n}" }, "type": AWS, "uri": "arn:aws:apigateway:us-east-1:kinesis:actionListStreams", "_embedded": { "integration:responses": { "_links": { "self": { "href": "restapisfugvjdxtriresources3kzxbg5sa2methodsGETintegrationresponses200", "name": "200", "title": "200" }, "integrationresponse:delete": { "href": "restapisfugvjdxtriresources3kzxbg5sa2methodsGETintegrationresponses200" }, "integrationresponse:update": { "href": "restapisfugvjdxtriresources3kzxbg5sa2methodsGETintegrationresponses200" } }, "responseParameters": { "method.response.header.Content-Type": "'applicationxml'" }, "responseTemplates": { "applicationjson": "$util.urlDecode("%3CkinesisStreams%3E%23foreach(%24stream%20in%20%24input.path(%27%24.StreamNames%27))%3Cstream%3E%3Cname%3E%24stream%3C%2Fname%3E%3C%2Fstream%3E%23end%3C%2FkinesisStreams%3E")" }, "statusCode": "200" } } }, "method:responses": { "_links": { "self": { "href": "restapisfugvjdxtriresources3kzxbg5sa2methodsGETresponses200", "name": "200", "title": "200" }, "methodresponse:delete": { "href": "restapisfugvjdxtriresources3kzxbg5sa2methodsGETresponses200" }, "methodresponse:update": { "href": "restapisfugvjdxtriresources3kzxbg5sa2methodsGETresponses200" } }, "responseModels": { "applicationjson": Empty }, "responseParameters": { "method.response.header.Content-Type": false }, "statusCode": "200" } } } In the example above, the response template for the 200 OK response maps the JSON output from the ListStreams action in the back end to an XML output. The mapping template is URL-encoded as %3CkinesisStreams%3E%23foreach(%24stream%20in%20%24input.path(%27%24.StreamNames%27))%3Cstream%3E%3Cname%3E%24stream%3C%2Fname%3E%3C%2Fstream%3E%23end%3C%2FkinesisStreams%3E@ and the output is decoded using the http://docs.aws.amazon.com/apigateway/latest/developerguide/api-gateway-mapping-template-reference.html#util-templat-reference util.urlDecode()> helper function.

MethodResponse , Integration , IntegrationResponse , Resource , Set up an API's method

See: method smart constructor.

Response Lenses

mMethodResponses :: Lens' Method (HashMap Text MethodResponse)

Gets a method response associated with a given HTTP status code. The collection of method responses are encapsulated in a key-value map, where the key is a response's HTTP status code and the value is a MethodResponse resource that specifies the response returned to the caller from the back end through the integration response. Example: Get a 200 OK response of a GET method Request @GET restapisuojnr9hd57resources0cjtchmethodsGETresponses200 HTTP1.1 Content-Type: applicationjson Host: apigateway.us-east-1.amazonaws.com Content-Length: 117 X-Amz-Date: 20160613T215008Z Authorization: AWS4-HMAC-SHA256 Credential={access_key_ID}20160613us-east-1apigatewayaws4_request, SignedHeaders=content-type;host;x-amz-date, Signature={sig4_hash} Response The successful response returns a 200 OK status code and a payload similar to the following: { "_links": { "curies": { "href": "http://docs.aws.amazon.com/apigateway/latest/developerguide/restapi-method-response-{rel}.html", "name": "methodresponse", "templated": true }, "self": { "href": "restapisuojnr9hd57resources0cjtchmethodsGETresponses200", "title": "200" }, "methodresponse:delete": { "href": "restapisuojnr9hd57resources0cjtchmethodsGETresponses200" }, "methodresponse:update": { "href": "restapisuojnr9hd57resources0cjtchmethodsGETresponses200" } }, "responseModels": { "application/json": Empty }, "responseParameters": { "method.response.header.operator": false, "method.response.header.operand_2": false, "method.response.header.operand_1": false }, "statusCode": "200" } @ AWS CLI

mHttpMethod :: Lens' Method (Maybe Text)

The method's HTTP verb.

mRequestValidatorId :: Lens' Method (Maybe Text)

The identifier of a RequestValidator for request validation.

mRequestModels :: Lens' Method (HashMap Text Text)

A key-value map specifying data schemas, represented by Model resources, (as the mapped value) of the request payloads of given content types (as the mapping key).

mRequestParameters :: Lens' Method (HashMap Text Bool)

A key-value map defining required or optional method request parameters that can be accepted by Amazon API Gateway. A key is a method request parameter name matching the pattern of method.request.{location}.{name} , where location is querystring , path , or header and name is a valid and unique parameter name. The value associated with the key is a Boolean flag indicating whether the parameter is required (true ) or optional (false ). The method request parameter names defined here are available in Integration to be mapped to integration request parameters or templates.

mAuthorizerId :: Lens' Method (Maybe Text)

The identifier of an Authorizer to use on this method. The authorizationType must be CUSTOM .

mOperationName :: Lens' Method (Maybe Text)

A human-friendly operation identifier for the method. For example, you can assign the operationName of ListPets for the GET /pets method in PetStore example.

mAuthorizationType :: Lens' Method (Maybe Text)

The method's authorization type. Valid values are NONE for open access, AWS_IAM for using AWS IAM permissions, CUSTOM for using a custom authorizer, or COGNITO_USER_POOLS for using a Cognito user pool.

mApiKeyRequired :: Lens' Method (Maybe Bool)

A boolean flag specifying whether a valid ApiKey is required to invoke this method.

mMethodIntegration :: Lens' Method (Maybe Integration)

Gets the method's integration responsible for passing the client-submitted request to the back end and performing necessary transformations to make the request compliant with the back end. Example: Request @GET restapisuojnr9hd57resources0cjtchmethodsGETintegration HTTP1.1 Content-Type: applicationjson Host: apigateway.us-east-1.amazonaws.com Content-Length: 117 X-Amz-Date: 20160613T213210Z Authorization: AWS4-HMAC-SHA256 Credential={access_key_ID}20160613us-east-1apigateway/aws4_request, SignedHeaders=content-type;host;x-amz-date, Signature={sig4_hash} Response The successful response returns a 200 OK status code and a payload similar to the following: { "_links": { "curies": [ { "href": "http://docs.aws.amazon.com/apigateway/latest/developerguide/restapi-integration-{rel}.html", "name": "integration", "templated": true }, { "href": "http://docs.aws.amazon.com/apigateway/latest/developerguide/restapi-integration-response-{rel}.html", "name": "integrationresponse", "templated": true } ], "self": { "href": "restapisuojnr9hd57resources0cjtchmethodsGETintegration" }, "integration:delete": { "href": "restapisuojnr9hd57resources0cjtchmethodsGETintegration" }, "integration:responses": { "href": "restapisuojnr9hd57resources0cjtchmethodsGETintegrationresponses200", "name": "200", "title": "200" }, "integration:update": { "href": "restapisuojnr9hd57resources0cjtchmethodsGETintegration" }, "integrationresponse:put": { "href": "restapisuojnr9hd57resources0cjtchmethodsGETintegrationresponses{status_code}", "templated": true } }, "cacheKeyParameters": [], "cacheNamespace": "0cjtch", "credentials": "arn:aws:iam::123456789012:roleapigAwsProxyRole", "httpMethod": POST, "passthroughBehavior": WHEN_NO_MATCH, "requestTemplates": { "applicationjson": "{n "a": "$input.params(operand1)",n "b": "$input.params(operand2)", n "op": "$input.params(operator)" n}" }, "type": AWS, "uri": "arn:aws:apigateway:us-west-2:lambda:path2015-03-31functionsarn:aws:lambda:us-west-2:123456789012:function:Calcinvocations", "_embedded": { "integration:responses": { "_links": { "self": { "href": "restapisuojnr9hd57resources0cjtchmethodsGETintegrationresponses200", "name": "200", "title": "200" }, "integrationresponse:delete": { "href": "restapisuojnr9hd57resources0cjtchmethodsGETintegrationresponses200" }, "integrationresponse:update": { "href": "restapisuojnr9hd57resources0cjtchmethodsGETintegrationresponses200" } }, "responseParameters": { "method.response.header.operator": "integration.response.body.op", "method.response.header.operand_2": "integration.response.body.b", "method.response.header.operand_1": "integration.response.body.a" }, "responseTemplates": { "applicationjson": "#set($res = $input.path($))n{n "result": "$res.a, $res.b, $res.op => $res.c",n "a" : "$res.a",n "b" : "$res.b",n "op" : "$res.op",n "c" : "$res.c"n}" }, "selectionPattern": "", "statusCode": "200" } } } @ AWS CLI