Network.AWS.APIGateway.GetMethodResponse

Description

Describes a MethodResponse resource.

Creating a Request

getMethodResponse

Arguments

:: Text
gmRestAPIId

-> Text
gmResourceId

-> Text
gmHttpMethod

-> Text
gmStatusCode

-> GetMethodResponse

Creates a value of GetMethodResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gmRestAPIId - The string identifier of the associated RestApi .

gmResourceId - The Resource identifier for the MethodResponse resource.

gmHttpMethod - The HTTP verb of the Method resource.

gmStatusCode - The status code for the MethodResponse resource.

data GetMethodResponse

Request to describe a MethodResponse resource.

See: getMethodResponse smart constructor.

Request Lenses

gmRestAPIId :: Lens' GetMethodResponse Text

The string identifier of the associated RestApi .

gmResourceId :: Lens' GetMethodResponse Text

The Resource identifier for the MethodResponse resource.

gmHttpMethod :: Lens' GetMethodResponse Text

The HTTP verb of the Method resource.

gmStatusCode :: Lens' GetMethodResponse Text

The status code for the MethodResponse resource.

Destructuring the Response

methodResponse :: MethodResponse

Creates a value of MethodResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mResponseModels - Specifies the Model resources used for the response's content-type. Response models are represented as a key/value map, with a content-type as the key and a Model name as the value.

mStatusCode - The method response's status code.

mResponseParameters - A key-value map specifying required or optional response parameters that Amazon API Gateway can send back to the caller. A key defines a method response header and the value specifies whether the associated method response header is required or not. The expression of the key must match the pattern method.response.header.{name} , where name is a valid and unique header name. Amazon API Gateway passes certain integration response data to the method response headers specified here according to the mapping you prescribe in the API's IntegrationResponse . The integration response data that can be mapped include an integration response header expressed in integration.response.header.{name} , a static value enclosed within a pair of single quotes (e.g., 'application/json' ), or a JSON expression from the back-end response payload in the form of integration.response.body.{JSON-expression} , where JSON-expression is a valid JSON expression without the > prefix.)

data MethodResponse

Represents a method response of a given HTTP status code returned to the client. The method response is passed from the back end through the associated integration response that can be transformed using a mapping template.

Example: A MethodResponse instance of an API Request The example request retrieves a MethodResponse of the 200 status code.

@GET restapisfugvjdxtriresources3kzxbg5sa2methodsGETresponses200 HTTP1.1 Content-Type: applicationjson Host: apigateway.us-east-1.amazonaws.com X-Amz-Date: 20160603T222952Z Authorization: AWS4-HMAC-SHA256 Credential={access_key_ID}20160603us-east-1apigatewayaws4_request, SignedHeaders=content-type;host;x-amz-date, Signature={sig4_hash} Response The successful response returns 200 OK@ status and a payload as follows:

@{ "_links": { "curies": { "href": "http://docs.aws.amazon.com/apigateway/latest/developerguide/restapi-method-response-{rel}.html", "name": "methodresponse", "templated": true }, "self": { "href": "restapisfugvjdxtriresources3kzxbg5sa2methodsGETresponses200", "title": "200" }, "methodresponse:delete": { "href": "restapisfugvjdxtriresources3kzxbg5sa2methodsGETresponses200" }, "methodresponse:update": { "href": "restapisfugvjdxtriresources3kzxbg5sa2methodsGETresponses200" } }, "responseModels": { "application/json": Empty }, "responseParameters": { "method.response.header.Content-Type": false }, "statusCode": "200" } @

Method , IntegrationResponse , Integration Creating an API

See: methodResponse smart constructor.

Response Lenses

mResponseModels :: Lens' MethodResponse (HashMap Text Text)

Specifies the Model resources used for the response's content-type. Response models are represented as a key/value map, with a content-type as the key and a Model name as the value.

mStatusCode :: Lens' MethodResponse (Maybe Text)

The method response's status code.

mResponseParameters :: Lens' MethodResponse (HashMap Text Bool)

A key-value map specifying required or optional response parameters that Amazon API Gateway can send back to the caller. A key defines a method response header and the value specifies whether the associated method response header is required or not. The expression of the key must match the pattern method.response.header.{name} , where name is a valid and unique header name. Amazon API Gateway passes certain integration response data to the method response headers specified here according to the mapping you prescribe in the API's IntegrationResponse . The integration response data that can be mapped include an integration response header expressed in integration.response.header.{name} , a static value enclosed within a pair of single quotes (e.g., 'application/json' ), or a JSON expression from the back-end response payload in the form of integration.response.body.{JSON-expression} , where JSON-expression is a valid JSON expression without the > prefix.)