Network.AWS.APIGateway.GetModel

Description

Describes an existing model defined for a RestApi resource.

Creating a Request

getModel

Arguments

:: Text
ggRestAPIId

-> Text
ggModelName

-> GetModel

Creates a value of GetModel with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ggFlatten - A query parameter of a Boolean value to resolve (true ) all external model references and returns a flattened model schema or not (false ) The default is false .

ggRestAPIId - The RestApi identifier under which the Model exists.

ggModelName - The name of the model as an identifier.

data GetModel

Request to list information about a model in an existing RestApi resource.

See: getModel smart constructor.

Request Lenses

ggFlatten :: Lens' GetModel (Maybe Bool)

A query parameter of a Boolean value to resolve (true ) all external model references and returns a flattened model schema or not (false ) The default is false .

ggRestAPIId :: Lens' GetModel Text

The RestApi identifier under which the Model exists.

ggModelName :: Lens' GetModel Text

The name of the model as an identifier.

Destructuring the Response

model :: Model

Creates a value of Model with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mSchema - The schema for the model. For application/json models, this should be JSON-schema draft v4 model. Do not include "*" characters in the description of any properties because such "*" characters may be interpreted as the closing marker for comments in some languages, such as Java or JavaScript, causing the installation of your API's SDK generated by API Gateway to fail.

mName - The name of the model.

mId - The identifier for the model resource.

mDescription - The description of the model.

mContentType - The content-type for the model.

data Model

Represents the data structure of a method's request or response payload.

A request model defines the data structure of the client-supplied request payload. A response model defines the data structure of the response payload returned by the back end. Although not required, models are useful for mapping payloads between the front end and back end.

A model is used for generating an API's SDK, validating the input request body, and creating a skeletal mapping template.

Method , MethodResponse , Models and Mappings

See: model smart constructor.

Response Lenses

mSchema :: Lens' Model (Maybe Text)

The schema for the model. For application/json models, this should be JSON-schema draft v4 model. Do not include "*" characters in the description of any properties because such "*" characters may be interpreted as the closing marker for comments in some languages, such as Java or JavaScript, causing the installation of your API's SDK generated by API Gateway to fail.

mName :: Lens' Model (Maybe Text)

The name of the model.

mId :: Lens' Model (Maybe Text)

The identifier for the model resource.

mDescription :: Lens' Model (Maybe Text)

The description of the model.

mContentType :: Lens' Model (Maybe Text)

The content-type for the model.