Network.AWS.APIGateway.GetRequestValidators

Description

Gets the RequestValidators collection of a given RestApi .

Creating a Request

getRequestValidators

Arguments

:: Text
grvRestAPIId

-> GetRequestValidators

Creates a value of GetRequestValidators with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

grvLimit - The maximum number of returned results per page.

grvPosition - The current pagination position in the paged result set.

grvRestAPIId - The string identifier of the associated RestApi .

data GetRequestValidators

Gets the RequestValidators collection of a given RestApi .

See: getRequestValidators smart constructor.

Request Lenses

grvLimit :: Lens' GetRequestValidators (Maybe Int)

The maximum number of returned results per page.

grvPosition :: Lens' GetRequestValidators (Maybe Text)

The current pagination position in the paged result set.

grvRestAPIId :: Lens' GetRequestValidators Text

The string identifier of the associated RestApi .

Destructuring the Response

getRequestValidatorsResponse

Arguments

:: Int
grvrsResponseStatus

-> GetRequestValidatorsResponse

Creates a value of GetRequestValidatorsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

grvrsItems - The current page of elements from this collection.

grvrsPosition - Undocumented member.

grvrsResponseStatus - -- | The response status code.

data GetRequestValidatorsResponse

A collection of RequestValidator resources of a given RestApi .

In Swagger, the RequestValidators of an API is defined by the x-amazon-apigateway-request-validators extension.

Enable Basic Request Validation in API Gateway

See: getRequestValidatorsResponse smart constructor.

Response Lenses

grvrsItems :: Lens' GetRequestValidatorsResponse [RequestValidator]

The current page of elements from this collection.

grvrsPosition :: Lens' GetRequestValidatorsResponse (Maybe Text)

Undocumented member.

grvrsResponseStatus :: Lens' GetRequestValidatorsResponse Int

- | The response status code.