Network.AWS.APIGateway.GetResources

Description

Lists information about a collection of Resource resources.

This operation returns paginated results.

Creating a Request

getResources

Arguments

:: Text
grsRestAPIId

-> GetResources

Creates a value of GetResources with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

grsEmbed - A query parameter used to retrieve the specified resources embedded in the returned Resources resource in the response. This embed parameter value is a list of comma-separated strings. Currently, the request supports only retrieval of the embedded Method resources this way. The query parameter value must be a single-valued list and contain the "methods" string. For example, GET restapis{restapi_id}/resources?embed=methods .

grsLimit - The maximum number of returned results per page. The value is 25 by default and could be between 1 - 500.

grsPosition - The current pagination position in the paged result set.

grsRestAPIId - The string identifier of the associated RestApi .

data GetResources

Request to list information about a collection of resources.

See: getResources smart constructor.

Request Lenses

grsEmbed :: Lens' GetResources [Text]

A query parameter used to retrieve the specified resources embedded in the returned Resources resource in the response. This embed parameter value is a list of comma-separated strings. Currently, the request supports only retrieval of the embedded Method resources this way. The query parameter value must be a single-valued list and contain the "methods" string. For example, GET restapis{restapi_id}/resources?embed=methods .

grsLimit :: Lens' GetResources (Maybe Int)

The maximum number of returned results per page. The value is 25 by default and could be between 1 - 500.

grsPosition :: Lens' GetResources (Maybe Text)

The current pagination position in the paged result set.

grsRestAPIId :: Lens' GetResources Text

The string identifier of the associated RestApi .

Destructuring the Response

getResourcesResponse

Arguments

:: Int
grrsResponseStatus

-> GetResourcesResponse

Creates a value of GetResourcesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

grrsItems - The current page of elements from this collection.

grrsPosition - Undocumented member.

grrsResponseStatus - -- | The response status code.

data GetResourcesResponse

Represents a collection of Resource resources.

Create an API

See: getResourcesResponse smart constructor.

Response Lenses

grrsItems :: Lens' GetResourcesResponse [Resource]

The current page of elements from this collection.

grrsPosition :: Lens' GetResourcesResponse (Maybe Text)

Undocumented member.

grrsResponseStatus :: Lens' GetResourcesResponse Int

- | The response status code.