Network.AWS.APIGateway.GetRestAPIs

Description

Lists the RestApis resources for your collection.

This operation returns paginated results.

Creating a Request

getRestAPIs :: GetRestAPIs

Creates a value of GetRestAPIs with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

graLimit - The maximum number of returned results per page. The value is 25 by default and could be between 1 - 500.

graPosition - The current pagination position in the paged result set.

data GetRestAPIs

The GET request to list existing RestApis defined for your collection.

See: getRestAPIs smart constructor.

Request Lenses

graLimit :: Lens' GetRestAPIs (Maybe Int)

The maximum number of returned results per page. The value is 25 by default and could be between 1 - 500.

graPosition :: Lens' GetRestAPIs (Maybe Text)

The current pagination position in the paged result set.

Destructuring the Response

getRestAPIsResponse

Arguments

:: Int
grarsResponseStatus

-> GetRestAPIsResponse

Creates a value of GetRestAPIsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

grarsItems - The current page of elements from this collection.

grarsPosition - Undocumented member.

grarsResponseStatus - -- | The response status code.

data GetRestAPIsResponse

Contains references to your APIs and links that guide you in how to interact with your collection. A collection offers a paginated view of your APIs.

Create an API

See: getRestAPIsResponse smart constructor.

Response Lenses

grarsItems :: Lens' GetRestAPIsResponse [RestAPI]

The current page of elements from this collection.

grarsPosition :: Lens' GetRestAPIsResponse (Maybe Text)

Undocumented member.

grarsResponseStatus :: Lens' GetRestAPIsResponse Int

- | The response status code.