Network.AWS.APIGateway.GetSDKTypes

Description

Undocumented operation.

Creating a Request

getSDKTypes :: GetSDKTypes

Creates a value of GetSDKTypes with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gstLimit - The maximum number of returned results per page.

gstPosition - The current pagination position in the paged result set.

data GetSDKTypes

Get the SdkTypes collection.

See: getSDKTypes smart constructor.

Request Lenses

gstLimit :: Lens' GetSDKTypes (Maybe Int)

The maximum number of returned results per page.

gstPosition :: Lens' GetSDKTypes (Maybe Text)

The current pagination position in the paged result set.

Destructuring the Response

getSDKTypesResponse

Arguments

:: Int
gstrsResponseStatus

-> GetSDKTypesResponse

Creates a value of GetSDKTypesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gstrsItems - The current page of elements from this collection.

gstrsPosition - Undocumented member.

gstrsResponseStatus - -- | The response status code.

data GetSDKTypesResponse

The collection of SdkType instances.

See: getSDKTypesResponse smart constructor.

Response Lenses

gstrsItems :: Lens' GetSDKTypesResponse [SDKType]

The current page of elements from this collection.

gstrsPosition :: Lens' GetSDKTypesResponse (Maybe Text)

Undocumented member.

gstrsResponseStatus :: Lens' GetSDKTypesResponse Int

- | The response status code.