Network.AWS.APIGateway.GetStage

Description

Gets information about a Stage resource.

Creating a Request

getStage

Arguments

:: Text
gssRestAPIId

-> Text
gssStageName

-> GetStage

Creates a value of GetStage with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gssRestAPIId - The string identifier of the associated RestApi .

gssStageName - The name of the Stage resource to get information about.

data GetStage

Requests Amazon API Gateway to get information about a Stage resource.

See: getStage smart constructor.

Request Lenses

gssRestAPIId :: Lens' GetStage Text

The string identifier of the associated RestApi .

gssStageName :: Lens' GetStage Text

The name of the Stage resource to get information about.

Destructuring the Response

stage :: Stage

Creates a value of Stage with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sDeploymentId - The identifier of the Deployment that the stage points to.

sVariables - A map that defines the stage variables for a Stage resource. Variable names can have alphanumeric and underscore characters, and the values must match [A-Za-z0-9-._~:/?#&=,]+ .

sDocumentationVersion - The version of the associated API documentation.

sClientCertificateId - The identifier of a client certificate for an API stage.

sCreatedDate - The timestamp when the stage was created.

sCacheClusterStatus - The status of the cache cluster for the stage, if enabled.

sMethodSettings - A map that defines the method settings for a Stage resource. Keys (designated as /{method_setting_key below) are method paths defined as {resource_path}/{http_method} for an individual method override, or ** for overriding all methods in the stage.

sLastUpdatedDate - The timestamp when the stage last updated.

sCacheClusterSize - The size of the cache cluster for the stage, if enabled.

sCacheClusterEnabled - Specifies whether a cache cluster is enabled for the stage.

sStageName - The name of the stage is the first path segment in the Uniform Resource Identifier (URI) of a call to Amazon API Gateway.

sDescription - The stage's description.

data Stage

Represents a unique identifier for a version of a deployed RestApi that is callable by users.

Deploy an API

See: stage smart constructor.

Response Lenses

sDeploymentId :: Lens' Stage (Maybe Text)

The identifier of the Deployment that the stage points to.

sVariables :: Lens' Stage (HashMap Text Text)

A map that defines the stage variables for a Stage resource. Variable names can have alphanumeric and underscore characters, and the values must match [A-Za-z0-9-._~:/?#&=,]+ .

sDocumentationVersion :: Lens' Stage (Maybe Text)

The version of the associated API documentation.

sClientCertificateId :: Lens' Stage (Maybe Text)

The identifier of a client certificate for an API stage.

sCreatedDate :: Lens' Stage (Maybe UTCTime)

The timestamp when the stage was created.

sCacheClusterStatus :: Lens' Stage (Maybe CacheClusterStatus)

The status of the cache cluster for the stage, if enabled.

sMethodSettings :: Lens' Stage (HashMap Text MethodSetting)

A map that defines the method settings for a Stage resource. Keys (designated as /{method_setting_key below) are method paths defined as {resource_path}/{http_method} for an individual method override, or ** for overriding all methods in the stage.

sLastUpdatedDate :: Lens' Stage (Maybe UTCTime)

The timestamp when the stage last updated.

sCacheClusterSize :: Lens' Stage (Maybe CacheClusterSize)

The size of the cache cluster for the stage, if enabled.

sCacheClusterEnabled :: Lens' Stage (Maybe Bool)

Specifies whether a cache cluster is enabled for the stage.

sStageName :: Lens' Stage (Maybe Text)

The name of the stage is the first path segment in the Uniform Resource Identifier (URI) of a call to Amazon API Gateway.

sDescription :: Lens' Stage (Maybe Text)

The stage's description.