Network.AWS.APIGateway.GetStages

Description

Gets information about one or more Stage resources.

Creating a Request

getStages

Arguments

:: Text
gsRestAPIId

-> GetStages

Creates a value of GetStages with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gsDeploymentId - The stages' deployment identifiers.

gsRestAPIId - The string identifier of the associated RestApi .

data GetStages

Requests Amazon API Gateway to get information about one or more Stage resources.

See: getStages smart constructor.

Request Lenses

gsDeploymentId :: Lens' GetStages (Maybe Text)

The stages' deployment identifiers.

gsRestAPIId :: Lens' GetStages Text

The string identifier of the associated RestApi .

Destructuring the Response

getStagesResponse

Arguments

:: Int
gsrsResponseStatus

-> GetStagesResponse

Creates a value of GetStagesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gsrsItem - The current page of elements from this collection.

gsrsResponseStatus - -- | The response status code.

data GetStagesResponse

A list of Stage resources that are associated with the ApiKey resource.

Deploying API in Stages

See: getStagesResponse smart constructor.

Response Lenses

gsrsItem :: Lens' GetStagesResponse [Stage]

The current page of elements from this collection.

gsrsResponseStatus :: Lens' GetStagesResponse Int

- | The response status code.