Network.AWS.APIGateway.GetUsage

Description

Gets the usage data of a usage plan in a specified time interval.

This operation returns paginated results.

Creating a Request

getUsage

Arguments

:: Text
guUsagePlanId

-> Text
guStartDate

-> Text
guEndDate

-> GetUsage

Creates a value of GetUsage with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

guKeyId - The Id of the API key associated with the resultant usage data.

guLimit - The maximum number of returned results per page.

guPosition - The current pagination position in the paged result set.

guUsagePlanId - The Id of the usage plan associated with the usage data.

guStartDate - The starting date (e.g., 2016-01-01) of the usage data.

guEndDate - The ending date (e.g., 2016-12-31) of the usage data.

data GetUsage

The GET request to get the usage data of a usage plan in a specified time interval.

See: getUsage smart constructor.

Request Lenses

guKeyId :: Lens' GetUsage (Maybe Text)

The Id of the API key associated with the resultant usage data.

guLimit :: Lens' GetUsage (Maybe Int)

The maximum number of returned results per page.

guPosition :: Lens' GetUsage (Maybe Text)

The current pagination position in the paged result set.

guUsagePlanId :: Lens' GetUsage Text

The Id of the usage plan associated with the usage data.

guStartDate :: Lens' GetUsage Text

The starting date (e.g., 2016-01-01) of the usage data.

guEndDate :: Lens' GetUsage Text

The ending date (e.g., 2016-12-31) of the usage data.

Destructuring the Response

usage :: Usage

Creates a value of Usage with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uUsagePlanId - The plan Id associated with this usage data.

uEndDate - The ending date of the usage data.

uItems - The usage data, as daily logs of used and remaining quotas, over the specified time interval indexed over the API keys in a usage plan. For example, {..., "values" : { "{api_key}" : [ [0, 100], [10, 90], [100, 10]]} , where {api_key} stands for an API key value and the daily log entry is of the format [used quota, remaining quota] .

uStartDate - The starting date of the usage data.

uPosition - Undocumented member.

data Usage

Represents the usage data of a usage plan.

Create and Use Usage Plans , Manage Usage in a Usage Plan

See: usage smart constructor.

Response Lenses

uUsagePlanId :: Lens' Usage (Maybe Text)

The plan Id associated with this usage data.

uEndDate :: Lens' Usage (Maybe Text)

The ending date of the usage data.

uItems :: Lens' Usage (HashMap Text [[Integer]])

The usage data, as daily logs of used and remaining quotas, over the specified time interval indexed over the API keys in a usage plan. For example, {..., "values" : { "{api_key}" : [ [0, 100], [10, 90], [100, 10]]} , where {api_key} stands for an API key value and the daily log entry is of the format [used quota, remaining quota] .

uStartDate :: Lens' Usage (Maybe Text)

The starting date of the usage data.

uPosition :: Lens' Usage (Maybe Text)

Undocumented member.