Network.AWS.APIGateway.GetUsagePlanKeys

Description

Gets all the usage plan keys representing the API keys added to a specified usage plan.

This operation returns paginated results.

Creating a Request

getUsagePlanKeys

Arguments

:: Text
gupkUsagePlanId

-> GetUsagePlanKeys

Creates a value of GetUsagePlanKeys with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gupkNameQuery - A query parameter specifying the name of the to-be-returned usage plan keys.

gupkLimit - The maximum number of returned results per page.

gupkPosition - The current pagination position in the paged result set.

gupkUsagePlanId - The Id of the UsagePlan resource representing the usage plan containing the to-be-retrieved UsagePlanKey resource representing a plan customer.

data GetUsagePlanKeys

The GET request to get all the usage plan keys representing the API keys added to a specified usage plan.

See: getUsagePlanKeys smart constructor.

Request Lenses

gupkNameQuery :: Lens' GetUsagePlanKeys (Maybe Text)

A query parameter specifying the name of the to-be-returned usage plan keys.

gupkLimit :: Lens' GetUsagePlanKeys (Maybe Int)

The maximum number of returned results per page.

gupkPosition :: Lens' GetUsagePlanKeys (Maybe Text)

The current pagination position in the paged result set.

gupkUsagePlanId :: Lens' GetUsagePlanKeys Text

The Id of the UsagePlan resource representing the usage plan containing the to-be-retrieved UsagePlanKey resource representing a plan customer.

Destructuring the Response

getUsagePlanKeysResponse

Arguments

:: Int
gupkrsResponseStatus

-> GetUsagePlanKeysResponse

Creates a value of GetUsagePlanKeysResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gupkrsItems - The current page of elements from this collection.

gupkrsPosition - Undocumented member.

gupkrsResponseStatus - -- | The response status code.

data GetUsagePlanKeysResponse

Represents the collection of usage plan keys added to usage plans for the associated API keys and, possibly, other types of keys.

Create and Use Usage Plans

See: getUsagePlanKeysResponse smart constructor.

Response Lenses

gupkrsItems :: Lens' GetUsagePlanKeysResponse [UsagePlanKey]

The current page of elements from this collection.

gupkrsPosition :: Lens' GetUsagePlanKeysResponse (Maybe Text)

Undocumented member.

gupkrsResponseStatus :: Lens' GetUsagePlanKeysResponse Int

- | The response status code.