Network.AWS.APIGateway.GetUsagePlans

Description

Gets all the usage plans of the caller's account.

This operation returns paginated results.

Creating a Request

getUsagePlans :: GetUsagePlans

Creates a value of GetUsagePlans with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gupKeyId - The identifier of the API key associated with the usage plans.

gupLimit - The maximum number of returned results per page.

gupPosition - The current pagination position in the paged result set.

data GetUsagePlans

The GET request to get all the usage plans of the caller's account.

See: getUsagePlans smart constructor.

Request Lenses

gupKeyId :: Lens' GetUsagePlans (Maybe Text)

The identifier of the API key associated with the usage plans.

gupLimit :: Lens' GetUsagePlans (Maybe Int)

The maximum number of returned results per page.

gupPosition :: Lens' GetUsagePlans (Maybe Text)

The current pagination position in the paged result set.

Destructuring the Response

getUsagePlansResponse

Arguments

:: Int
guprsResponseStatus

-> GetUsagePlansResponse

Creates a value of GetUsagePlansResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

guprsItems - The current page of elements from this collection.

guprsPosition - Undocumented member.

guprsResponseStatus - -- | The response status code.

data GetUsagePlansResponse

Represents a collection of usage plans for an AWS account.

Create and Use Usage Plans

See: getUsagePlansResponse smart constructor.

Response Lenses

guprsItems :: Lens' GetUsagePlansResponse [UsagePlan]

The current page of elements from this collection.

guprsPosition :: Lens' GetUsagePlansResponse (Maybe Text)

Undocumented member.

guprsResponseStatus :: Lens' GetUsagePlansResponse Int

- | The response status code.