Network.AWS.APIGateway.ImportDocumentationParts

Description

Undocumented operation.

Creating a Request

importDocumentationParts

Arguments

:: Text
idpRestAPIId

-> HashMap Text Value
idpBody

-> ImportDocumentationParts

Creates a value of ImportDocumentationParts with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

idpMode - A query parameter to indicate whether to overwrite (OVERWRITE ) any existing DocumentationParts definition or to merge (MERGE ) the new definition into the existing one. The default value is MERGE .

idpFailOnWarnings - A query parameter to specify whether to rollback the documentation importation (true ) or not (false ) when a warning is encountered. The default value is false .

idpRestAPIId - [Required] The string identifier of the associated RestApi .

idpBody - [Required] Raw byte array representing the to-be-imported documentation parts. To import from a Swagger file, this is a JSON object.

data ImportDocumentationParts

Import documentation parts from an external (e.g., Swagger) definition file.

See: importDocumentationParts smart constructor.

Request Lenses

idpMode :: Lens' ImportDocumentationParts (Maybe PutMode)

A query parameter to indicate whether to overwrite (OVERWRITE ) any existing DocumentationParts definition or to merge (MERGE ) the new definition into the existing one. The default value is MERGE .

idpFailOnWarnings :: Lens' ImportDocumentationParts (Maybe Bool)

A query parameter to specify whether to rollback the documentation importation (true ) or not (false ) when a warning is encountered. The default value is false .

idpRestAPIId :: Lens' ImportDocumentationParts Text

Required

The string identifier of the associated RestApi .

idpBody :: Lens' ImportDocumentationParts (HashMap Text Value)

Required

Raw byte array representing the to-be-imported documentation parts. To import from a Swagger file, this is a JSON object.

Destructuring the Response

importDocumentationPartsResponse

Arguments

:: Int
idprsResponseStatus

-> ImportDocumentationPartsResponse

Creates a value of ImportDocumentationPartsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

idprsIds - A list of the returned documentation part identifiers.

idprsWarnings - A list of warning messages reported during import of documentation parts.

idprsResponseStatus - -- | The response status code.

data ImportDocumentationPartsResponse

A collection of the imported DocumentationPart identifiers.

This is used to return the result when documentation parts in an external (e.g., Swagger) file are imported into Amazon API GatewayDocumenting an API , documentationpart:import , DocumentationPart

See: importDocumentationPartsResponse smart constructor.

Response Lenses

idprsIds :: Lens' ImportDocumentationPartsResponse [Text]

A list of the returned documentation part identifiers.

idprsWarnings :: Lens' ImportDocumentationPartsResponse [Text]

A list of warning messages reported during import of documentation parts.

idprsResponseStatus :: Lens' ImportDocumentationPartsResponse Int

- | The response status code.