Network.AWS.APIGateway.ImportRestAPI

Description

A feature of the Amazon API Gateway control service for creating a new API from an external API definition file.

Creating a Request

importRestAPI

Arguments

:: HashMap Text Value
iraBody

-> ImportRestAPI

Creates a value of ImportRestAPI with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

iraFailOnWarnings - A query parameter to indicate whether to rollback the API creation (true ) or not (false ) when a warning is encountered. The default value is false .

iraParameters - Custom header parameters as part of the request. For example, to exclude DocumentationParts from an imported API, set ignore=documentation as a parameters value, as in the AWS CLI command of aws apigateway import-rest-api --parameters ignore=documentation --body 'file://pathto/imported-api-body.json .

iraBody - The POST request body containing external API definitions. Currently, only Swagger definition JSON files are supported. The maximum size of the API definition file is 2MB.

data ImportRestAPI

A POST request to import an API to Amazon API Gateway using an input of an API definition file.

See: importRestAPI smart constructor.

Request Lenses

iraFailOnWarnings :: Lens' ImportRestAPI (Maybe Bool)

A query parameter to indicate whether to rollback the API creation (true ) or not (false ) when a warning is encountered. The default value is false .

iraParameters :: Lens' ImportRestAPI (HashMap Text Text)

Custom header parameters as part of the request. For example, to exclude DocumentationParts from an imported API, set ignore=documentation as a parameters value, as in the AWS CLI command of aws apigateway import-rest-api --parameters ignore=documentation --body 'file://pathto/imported-api-body.json .

iraBody :: Lens' ImportRestAPI (HashMap Text Value)

The POST request body containing external API definitions. Currently, only Swagger definition JSON files are supported. The maximum size of the API definition file is 2MB.

Destructuring the Response

restAPI :: RestAPI

Creates a value of RestAPI with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

raBinaryMediaTypes - The list of binary media types supported by the RestApi . By default, the RestApi supports only UTF-8-encoded text payloads.

raWarnings - The warning messages reported when failonwarnings is turned on during API import.

raCreatedDate - The timestamp when the API was created.

raName - The API's name.

raVersion - A version identifier for the API.

raId - The API's identifier. This identifier is unique across all of your APIs in Amazon API Gateway.

raDescription - The API's description.

data RestAPI

Represents a REST API.

Create an API

See: restAPI smart constructor.

Response Lenses

raBinaryMediaTypes :: Lens' RestAPI [Text]

The list of binary media types supported by the RestApi . By default, the RestApi supports only UTF-8-encoded text payloads.

raWarnings :: Lens' RestAPI [Text]

The warning messages reported when failonwarnings is turned on during API import.

raCreatedDate :: Lens' RestAPI (Maybe UTCTime)

The timestamp when the API was created.

raName :: Lens' RestAPI (Maybe Text)

The API's name.

raVersion :: Lens' RestAPI (Maybe Text)

A version identifier for the API.

raId :: Lens' RestAPI (Maybe Text)

The API's identifier. This identifier is unique across all of your APIs in Amazon API Gateway.

raDescription :: Lens' RestAPI (Maybe Text)

The API's description.