Network.AWS.APIGateway.TestInvokeAuthorizer

Description

Simulate the execution of an Authorizer in your RestApi with headers, parameters, and an incoming request body.

Enable custom authorizers

Creating a Request

testInvokeAuthorizer

Arguments

:: Text
tiaRestAPIId

-> Text
tiaAuthorizerId

-> TestInvokeAuthorizer

Creates a value of TestInvokeAuthorizer with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tiaPathWithQueryString - [Optional] The URI path, including query string, of the simulated invocation request. Use this to specify path parameters and query string parameters.

tiaBody - [Optional] The simulated request body of an incoming invocation request.

tiaAdditionalContext - [Optional] A key-value map of additional context variables.

tiaStageVariables - A key-value map of stage variables to simulate an invocation on a deployed Stage .

tiaHeaders - [Required] A key-value map of headers to simulate an incoming invocation request. This is where the incoming authorization token, or identity source, should be specified.

tiaRestAPIId - The string identifier of the associated RestApi .

tiaAuthorizerId - Specifies a test invoke authorizer request's Authorizer ID.

data TestInvokeAuthorizer

Make a request to simulate the execution of an Authorizer .

See: testInvokeAuthorizer smart constructor.

Request Lenses

tiaPathWithQueryString :: Lens' TestInvokeAuthorizer (Maybe Text)

Optional

The URI path, including query string, of the simulated invocation request. Use this to specify path parameters and query string parameters.

tiaBody :: Lens' TestInvokeAuthorizer (Maybe Text)

Optional

The simulated request body of an incoming invocation request.

tiaAdditionalContext :: Lens' TestInvokeAuthorizer (HashMap Text Text)

Optional

A key-value map of additional context variables.

tiaStageVariables :: Lens' TestInvokeAuthorizer (HashMap Text Text)

A key-value map of stage variables to simulate an invocation on a deployed Stage .

tiaHeaders :: Lens' TestInvokeAuthorizer (HashMap Text Text)

Required

A key-value map of headers to simulate an incoming invocation request. This is where the incoming authorization token, or identity source, should be specified.

tiaRestAPIId :: Lens' TestInvokeAuthorizer Text

The string identifier of the associated RestApi .

tiaAuthorizerId :: Lens' TestInvokeAuthorizer Text

Specifies a test invoke authorizer request's Authorizer ID.

Destructuring the Response

testInvokeAuthorizerResponse

Arguments

:: Int
tiarsResponseStatus

-> TestInvokeAuthorizerResponse

Creates a value of TestInvokeAuthorizerResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tiarsLog - The Amazon API Gateway execution log for the test authorizer request.

tiarsPrincipalId - The principal identity returned by the Authorizer

tiarsLatency - The execution latency of the test authorizer request.

tiarsAuthorization - Undocumented member.

tiarsClaims - The open identity claims , with any supported custom attributes, returned from the Cognito Your User Pool configured for the API.

tiarsClientStatus - The HTTP status code that the client would have received. Value is 0 if the authorizer succeeded.

tiarsPolicy - The JSON policy document returned by the Authorizer

tiarsResponseStatus - -- | The response status code.

data TestInvokeAuthorizerResponse

Represents the response of the test invoke request for a custom Authorizer

See: testInvokeAuthorizerResponse smart constructor.

Response Lenses

tiarsLog :: Lens' TestInvokeAuthorizerResponse (Maybe Text)

The Amazon API Gateway execution log for the test authorizer request.

tiarsPrincipalId :: Lens' TestInvokeAuthorizerResponse (Maybe Text)

The principal identity returned by the Authorizer

tiarsLatency :: Lens' TestInvokeAuthorizerResponse (Maybe Integer)

The execution latency of the test authorizer request.

tiarsAuthorization :: Lens' TestInvokeAuthorizerResponse (HashMap Text [Text])

Undocumented member.

tiarsClaims :: Lens' TestInvokeAuthorizerResponse (HashMap Text Text)

The open identity claims , with any supported custom attributes, returned from the Cognito Your User Pool configured for the API.

tiarsClientStatus :: Lens' TestInvokeAuthorizerResponse (Maybe Int)

The HTTP status code that the client would have received. Value is 0 if the authorizer succeeded.

tiarsPolicy :: Lens' TestInvokeAuthorizerResponse (Maybe Text)

The JSON policy document returned by the Authorizer

tiarsResponseStatus :: Lens' TestInvokeAuthorizerResponse Int

- | The response status code.