Network.AWS.APIGateway.TestInvokeMethod

Description

Simulate the execution of a Method in your RestApi with headers, parameters, and an incoming request body.

Creating a Request

testInvokeMethod

Arguments

:: Text
timRestAPIId

-> Text
timResourceId

-> Text
timHttpMethod

-> TestInvokeMethod

Creates a value of TestInvokeMethod with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

timPathWithQueryString - The URI path, including query string, of the simulated invocation request. Use this to specify path parameters and query string parameters.

timBody - The simulated request body of an incoming invocation request.

timClientCertificateId - A ClientCertificate identifier to use in the test invocation. API Gateway will use the certificate when making the HTTPS request to the defined back-end endpoint.

timStageVariables - A key-value map of stage variables to simulate an invocation on a deployed Stage .

timHeaders - A key-value map of headers to simulate an incoming invocation request.

timRestAPIId - The string identifier of the associated RestApi .

timResourceId - Specifies a test invoke method request's resource ID.

timHttpMethod - Specifies a test invoke method request's HTTP method.

data TestInvokeMethod

Make a request to simulate the execution of a Method .

See: testInvokeMethod smart constructor.

Request Lenses

timPathWithQueryString :: Lens' TestInvokeMethod (Maybe Text)

The URI path, including query string, of the simulated invocation request. Use this to specify path parameters and query string parameters.

timBody :: Lens' TestInvokeMethod (Maybe Text)

The simulated request body of an incoming invocation request.

timClientCertificateId :: Lens' TestInvokeMethod (Maybe Text)

A ClientCertificate identifier to use in the test invocation. API Gateway will use the certificate when making the HTTPS request to the defined back-end endpoint.

timStageVariables :: Lens' TestInvokeMethod (HashMap Text Text)

A key-value map of stage variables to simulate an invocation on a deployed Stage .

timHeaders :: Lens' TestInvokeMethod (HashMap Text Text)

A key-value map of headers to simulate an incoming invocation request.

timRestAPIId :: Lens' TestInvokeMethod Text

The string identifier of the associated RestApi .

timResourceId :: Lens' TestInvokeMethod Text

Specifies a test invoke method request's resource ID.

timHttpMethod :: Lens' TestInvokeMethod Text

Specifies a test invoke method request's HTTP method.

Destructuring the Response

testInvokeMethodResponse

Arguments

:: Int
timrsResponseStatus

-> TestInvokeMethodResponse

Creates a value of TestInvokeMethodResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

timrsLog - The Amazon API Gateway execution log for the test invoke request.

timrsStatus - The HTTP status code.

timrsBody - The body of the HTTP response.

timrsLatency - The execution latency of the test invoke request.

timrsHeaders - The headers of the HTTP response.

timrsResponseStatus - -- | The response status code.

data TestInvokeMethodResponse

Represents the response of the test invoke request in the HTTP method.

Test API using the API Gateway console

See: testInvokeMethodResponse smart constructor.

Response Lenses

timrsLog :: Lens' TestInvokeMethodResponse (Maybe Text)

The Amazon API Gateway execution log for the test invoke request.

timrsStatus :: Lens' TestInvokeMethodResponse (Maybe Int)

The HTTP status code.

timrsBody :: Lens' TestInvokeMethodResponse (Maybe Text)

The body of the HTTP response.

timrsLatency :: Lens' TestInvokeMethodResponse (Maybe Integer)

The execution latency of the test invoke request.

timrsHeaders :: Lens' TestInvokeMethodResponse (HashMap Text Text)

The headers of the HTTP response.

timrsResponseStatus :: Lens' TestInvokeMethodResponse Int

- | The response status code.