Network.AWS.APIGateway.UpdateAccount

Description

Changes information about the current Account resource.

Creating a Request

updateAccount :: UpdateAccount

Creates a value of UpdateAccount with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uaPatchOperations - A list of update operations to be applied to the specified resource and in the order specified in this list.

data UpdateAccount

Requests Amazon API Gateway to change information about the current Account resource.

See: updateAccount smart constructor.

Request Lenses

uaPatchOperations :: Lens' UpdateAccount [PatchOperation]

A list of update operations to be applied to the specified resource and in the order specified in this list.

Destructuring the Response

account :: Account

Creates a value of Account with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aApiKeyVersion - The version of the API keys used for the account.

aCloudwatchRoleARN - The ARN of an Amazon CloudWatch role for the current Account .

aFeatures - A list of features supported for the account. When usage plans are enabled, the features list will include an entry of UsagePlans .

aThrottleSettings - Specifies the API request limits configured for the current Account .

data Account

Represents an AWS account that is associated with Amazon API Gateway.

To view the account info, call GET on this resource.

Error Codes The following exception may be thrown when the request fails.

UnauthorizedException * NotFoundException * TooManyRequestsException

For detailed error code information, including the corresponding HTTP Status Codes, see API Gateway Error Codes

Example: Get the information about an account. Request @GET account HTTP1.1 Content-Type: applicationjson Host: apigateway.us-east-1.amazonaws.com X-Amz-Date: 20160531T184618Z Authorization: AWS4-HMAC-SHA256 Credential={access_key_ID}us-east-1apigatewayaws4_request, SignedHeaders=content-type;host;x-amz-date, Signature={sig4_hash} Response The successful response returns a 200 OK@ status code and a payload similar to the following:

@{ "_links": { "curies": { "href": "http://docs.aws.amazon.com/apigateway/latest/developerguide/account-apigateway-{rel}.html", "name": "account", "templated": true }, "self": { "href": "account" }, "account:update": { "href": "account" } }, "cloudwatchRoleArn": "arn:aws:iam::123456789012:role/apigAwsProxyRole", "throttleSettings": { "rateLimit": 500, "burstLimit": 1000 } } @ In addition to making the REST API call directly, you can use the AWS CLI and an AWS SDK to access this resource.

API Gateway Limits Developer Guide , AWS CLI

See: account smart constructor.

Response Lenses

aApiKeyVersion :: Lens' Account (Maybe Text)

The version of the API keys used for the account.

aCloudwatchRoleARN :: Lens' Account (Maybe Text)

The ARN of an Amazon CloudWatch role for the current Account .

aFeatures :: Lens' Account [Text]

A list of features supported for the account. When usage plans are enabled, the features list will include an entry of UsagePlans .

aThrottleSettings :: Lens' Account (Maybe ThrottleSettings)

Specifies the API request limits configured for the current Account .