Network.AWS.APIGateway.UpdateAuthorizer

Description

Updates an existing Authorizer resource.

AWS CLI

Creating a Request

updateAuthorizer

Arguments

:: Text
uaaRestAPIId

-> Text
uaaAuthorizerId

-> UpdateAuthorizer

Creates a value of UpdateAuthorizer with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uaaPatchOperations - A list of update operations to be applied to the specified resource and in the order specified in this list.

uaaRestAPIId - The string identifier of the associated RestApi .

uaaAuthorizerId - The identifier of the Authorizer resource.

data UpdateAuthorizer

Request to update an existing Authorizer resource.

See: updateAuthorizer smart constructor.

Request Lenses

uaaPatchOperations :: Lens' UpdateAuthorizer [PatchOperation]

A list of update operations to be applied to the specified resource and in the order specified in this list.

uaaRestAPIId :: Lens' UpdateAuthorizer Text

The string identifier of the associated RestApi .

uaaAuthorizerId :: Lens' UpdateAuthorizer Text

The identifier of the Authorizer resource.

Destructuring the Response

authorizer :: Authorizer

Creates a value of Authorizer with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aAuthorizerURI - [Required] Specifies the authorizer's Uniform Resource Identifier (URI). For TOKEN authorizers, this must be a well-formed Lambda function URI, for example, arn:aws:apigateway:us-west-2:lambda:path2015-03-31functionsarn:aws:lambda:us-west-2:{account_id}:function:{lambda_function_name}invocations . In general, the URI has this form arn:aws:apigateway:{region}:lambda:path/{service_api} , where {region} is the same as the region hosting the Lambda function, path indicates that the remaining substring in the URI should be treated as the path to the resource, including the initial / . For Lambda functions, this is usually of the form 2015-03-31functions[FunctionARN]invocations.

aIdentityValidationExpression - A validation expression for the incoming identity. For TOKEN authorizers, this value should be a regular expression. The incoming token from the client is matched against this expression, and will proceed if the token matches. If the token doesn't match, the client receives a 401 Unauthorized response.

aProviderARNs - A list of the provider ARNs of the authorizer. For an TOKEN authorizer, this is not defined. For authorizers of the COGNITO_USER_POOLS type, each element corresponds to a user pool ARN of this format: arn:aws:cognito-idp:{region}:{account_id}:userpool/{user_pool_id} .

aName - [Required] The name of the authorizer.

aId - The identifier for the authorizer resource.

aAuthorizerResultTtlInSeconds - The TTL in seconds of cached authorizer results. If greater than 0, API Gateway will cache authorizer responses. If this field is not set, the default value is 300. The maximum value is 3600, or 1 hour.

aAuthType - Optional customer-defined field, used in Swagger imports/exports. Has no functional impact.

aType - [Required] The type of the authorizer. Currently, the valid type is TOKEN for a Lambda function or COGNITO_USER_POOLS for an Amazon Cognito user pool.

aIdentitySource - [Required] The source of the identity in an incoming request. For a TOKEN authorizer, this value is a mapping expression with the same syntax as integration parameter mappings. The only valid source for tokens is header, so the expression should match 'method.request.header.[headerName]'. The value of the header '[headerName]' will be interpreted as the incoming token. For COGNITO_USER_POOLS authorizers, this property is used.

aAuthorizerCredentials - Specifies the credentials required for the authorizer, if any. Two options are available. To specify an IAM role for Amazon API Gateway to assume, use the role's Amazon Resource Name (ARN). To use resource-based permissions on the Lambda function, specify null.

data Authorizer

Represents an authorization layer for methods. If enabled on a method, API Gateway will activate the authorizer when a client calls the method.

Enable custom authorization

See: authorizer smart constructor.

Response Lenses

aAuthorizerURI :: Lens' Authorizer (Maybe Text)

Required

Specifies the authorizer's Uniform Resource Identifier (URI). For TOKEN authorizers, this must be a well-formed Lambda function URI, for example, arn:aws:apigateway:us-west-2:lambda:path2015-03-31functionsarn:aws:lambda:us-west-2:{account_id}:function:{lambda_function_name}invocations . In general, the URI has this form arn:aws:apigateway:{region}:lambda:path/{service_api} , where {region} is the same as the region hosting the Lambda function, path indicates that the remaining substring in the URI should be treated as the path to the resource, including the initial / . For Lambda functions, this is usually of the form 2015-03-31functions[FunctionARN]invocations.

aIdentityValidationExpression :: Lens' Authorizer (Maybe Text)

A validation expression for the incoming identity. For TOKEN authorizers, this value should be a regular expression. The incoming token from the client is matched against this expression, and will proceed if the token matches. If the token doesn't match, the client receives a 401 Unauthorized response.

aProviderARNs :: Lens' Authorizer [Text]

A list of the provider ARNs of the authorizer. For an TOKEN authorizer, this is not defined. For authorizers of the COGNITO_USER_POOLS type, each element corresponds to a user pool ARN of this format: arn:aws:cognito-idp:{region}:{account_id}:userpool/{user_pool_id} .

aName :: Lens' Authorizer (Maybe Text)

Required

The name of the authorizer.

aId :: Lens' Authorizer (Maybe Text)

The identifier for the authorizer resource.

aAuthorizerResultTtlInSeconds :: Lens' Authorizer (Maybe Int)

The TTL in seconds of cached authorizer results. If greater than 0, API Gateway will cache authorizer responses. If this field is not set, the default value is 300. The maximum value is 3600, or 1 hour.

aAuthType :: Lens' Authorizer (Maybe Text)

Optional customer-defined field, used in Swagger imports/exports. Has no functional impact.

aType :: Lens' Authorizer (Maybe AuthorizerType)

Required

The type of the authorizer. Currently, the valid type is TOKEN for a Lambda function or COGNITO_USER_POOLS for an Amazon Cognito user pool.

aIdentitySource :: Lens' Authorizer (Maybe Text)

Required

The source of the identity in an incoming request. For a TOKEN authorizer, this value is a mapping expression with the same syntax as integration parameter mappings. The only valid source for tokens is header, so the expression should match 'method.request.header.[headerName]'. The value of the header '[headerName]' will be interpreted as the incoming token. For COGNITO_USER_POOLS authorizers, this property is used.

aAuthorizerCredentials :: Lens' Authorizer (Maybe Text)

Specifies the credentials required for the authorizer, if any. Two options are available. To specify an IAM role for Amazon API Gateway to assume, use the role's Amazon Resource Name (ARN). To use resource-based permissions on the Lambda function, specify null.