Network.AWS.APIGateway.UpdateBasePathMapping

Description

Changes information about the BasePathMapping resource.

Creating a Request

updateBasePathMapping

Arguments

:: Text
ubpmDomainName

-> Text
ubpmBasePath

-> UpdateBasePathMapping

Creates a value of UpdateBasePathMapping with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ubpmPatchOperations - A list of update operations to be applied to the specified resource and in the order specified in this list.

ubpmDomainName - The domain name of the BasePathMapping resource to change.

ubpmBasePath - The base path of the BasePathMapping resource to change.

data UpdateBasePathMapping

A request to change information about the BasePathMapping resource.

See: updateBasePathMapping smart constructor.

Request Lenses

ubpmPatchOperations :: Lens' UpdateBasePathMapping [PatchOperation]

A list of update operations to be applied to the specified resource and in the order specified in this list.

ubpmDomainName :: Lens' UpdateBasePathMapping Text

The domain name of the BasePathMapping resource to change.

ubpmBasePath :: Lens' UpdateBasePathMapping Text

The base path of the BasePathMapping resource to change.

Destructuring the Response

basePathMapping :: BasePathMapping

Creates a value of BasePathMapping with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

bpmStage - The name of the associated stage.

bpmBasePath - The base path name that callers of the API must provide as part of the URL after the domain name.

bpmRestAPIId - The string identifier of the associated RestApi .

data BasePathMapping

Represents the base path that callers of the API must provide as part of the URL after the domain name.

A custom domain name plus a BasePathMapping specification identifies a deployed RestApi in a given stage of the owner Account .Use Custom Domain Names

See: basePathMapping smart constructor.

Response Lenses

bpmStage :: Lens' BasePathMapping (Maybe Text)

The name of the associated stage.

bpmBasePath :: Lens' BasePathMapping (Maybe Text)

The base path name that callers of the API must provide as part of the URL after the domain name.

bpmRestAPIId :: Lens' BasePathMapping (Maybe Text)

The string identifier of the associated RestApi .