Network.AWS.APIGateway.UpdateIntegrationResponse

Description

Represents an update integration response.

Creating a Request

updateIntegrationResponse

Arguments

:: Text
uiRestAPIId

-> Text
uiResourceId

-> Text
uiHttpMethod

-> Text
uiStatusCode

-> UpdateIntegrationResponse

Creates a value of UpdateIntegrationResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uiPatchOperations - A list of update operations to be applied to the specified resource and in the order specified in this list.

uiRestAPIId - The string identifier of the associated RestApi .

uiResourceId - Specifies an update integration response request's resource identifier.

uiHttpMethod - Specifies an update integration response request's HTTP method.

uiStatusCode - Specifies an update integration response request's status code.

data UpdateIntegrationResponse

Represents an update integration response request.

See: updateIntegrationResponse smart constructor.

Request Lenses

uiPatchOperations :: Lens' UpdateIntegrationResponse [PatchOperation]

A list of update operations to be applied to the specified resource and in the order specified in this list.

uiRestAPIId :: Lens' UpdateIntegrationResponse Text

The string identifier of the associated RestApi .

uiResourceId :: Lens' UpdateIntegrationResponse Text

Specifies an update integration response request's resource identifier.

uiHttpMethod :: Lens' UpdateIntegrationResponse Text

Specifies an update integration response request's HTTP method.

uiStatusCode :: Lens' UpdateIntegrationResponse Text

Specifies an update integration response request's status code.

Destructuring the Response

integrationResponse :: IntegrationResponse

Creates a value of IntegrationResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

intContentHandling - Specifies how to handle response payload content type conversions. Supported values are CONVERT_TO_BINARY and CONVERT_TO_TEXT , with the following behaviors: * CONVERT_TO_BINARY : Converts a response payload from a Base64-encoded string to the corresponding binary blob. * CONVERT_TO_TEXT : Converts a response payload from a binary blob to a Base64-encoded string. If this property is not defined, the response payload will be passed through from the integration response to the method response without modification.

intResponseTemplates - Specifies the templates used to transform the integration response body. Response templates are represented as a key/value map, with a content-type as the key and a template as the value.

intSelectionPattern - Specifies the regular expression (regex) pattern used to choose an integration response based on the response from the back end. For example, if the success response returns nothing and the error response returns some string, you could use the .+ regex to match error response. However, make sure that the error response does not contain any newline (n ) character in such cases. If the back end is an AWS Lambda function, the AWS Lambda function error header is matched. For all other HTTP and AWS back ends, the HTTP status code is matched.

intStatusCode - Specifies the status code that is used to map the integration response to an existing MethodResponse .

intResponseParameters - A key-value map specifying response parameters that are passed to the method response from the back end. The key is a method response header parameter name and the mapped value is an integration response header value, a static value enclosed within a pair of single quotes, or a JSON expression from the integration response body. The mapping key must match the pattern of method.response.header.{name} , where name is a valid and unique header name. The mapped non-static value must match the pattern of integration.response.header.{name} or integration.response.body.{JSON-expression} , where name is a valid and unique response header name and JSON-expression is a valid JSON expression without the > prefix.

data IntegrationResponse

Represents an integration response. The status code must map to an existing MethodResponse , and parameters and templates can be used to transform the back-end response.

Creating an API

See: integrationResponse smart constructor.

Response Lenses

intContentHandling :: Lens' IntegrationResponse (Maybe ContentHandlingStrategy)

Specifies how to handle response payload content type conversions. Supported values are CONVERT_TO_BINARY and CONVERT_TO_TEXT , with the following behaviors: * CONVERT_TO_BINARY : Converts a response payload from a Base64-encoded string to the corresponding binary blob. * CONVERT_TO_TEXT : Converts a response payload from a binary blob to a Base64-encoded string. If this property is not defined, the response payload will be passed through from the integration response to the method response without modification.

intResponseTemplates :: Lens' IntegrationResponse (HashMap Text Text)

Specifies the templates used to transform the integration response body. Response templates are represented as a key/value map, with a content-type as the key and a template as the value.

intSelectionPattern :: Lens' IntegrationResponse (Maybe Text)

Specifies the regular expression (regex) pattern used to choose an integration response based on the response from the back end. For example, if the success response returns nothing and the error response returns some string, you could use the .+ regex to match error response. However, make sure that the error response does not contain any newline (n ) character in such cases. If the back end is an AWS Lambda function, the AWS Lambda function error header is matched. For all other HTTP and AWS back ends, the HTTP status code is matched.

intStatusCode :: Lens' IntegrationResponse (Maybe Text)

Specifies the status code that is used to map the integration response to an existing MethodResponse .

intResponseParameters :: Lens' IntegrationResponse (HashMap Text Text)

A key-value map specifying response parameters that are passed to the method response from the back end. The key is a method response header parameter name and the mapped value is an integration response header value, a static value enclosed within a pair of single quotes, or a JSON expression from the integration response body. The mapping key must match the pattern of method.response.header.{name} , where name is a valid and unique header name. The mapped non-static value must match the pattern of integration.response.header.{name} or integration.response.body.{JSON-expression} , where name is a valid and unique response header name and JSON-expression is a valid JSON expression without the > prefix.