Network.AWS.APIGateway.UpdateUsagePlan

Description

Updates a usage plan of a given plan Id.

Creating a Request

updateUsagePlan

Arguments

:: Text
uupUsagePlanId

-> UpdateUsagePlan

Creates a value of UpdateUsagePlan with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uupPatchOperations - A list of update operations to be applied to the specified resource and in the order specified in this list.

uupUsagePlanId - The Id of the to-be-updated usage plan.

data UpdateUsagePlan

The PATCH request to update a usage plan of a given plan Id.

See: updateUsagePlan smart constructor.

Request Lenses

uupPatchOperations :: Lens' UpdateUsagePlan [PatchOperation]

A list of update operations to be applied to the specified resource and in the order specified in this list.

uupUsagePlanId :: Lens' UpdateUsagePlan Text

The Id of the to-be-updated usage plan.

Destructuring the Response

usagePlan :: UsagePlan

Creates a value of UsagePlan with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

upApiStages - The associated API stages of a usage plan.

upName - The name of a usage plan.

upId - The identifier of a UsagePlan resource.

upThrottle - The request throttle limits of a usage plan.

upQuota - The maximum number of permitted requests per a given unit time interval.

upDescription - The description of a usage plan.

upProductCode - The AWS Markeplace product identifier to associate with the usage plan as a SaaS product on AWS Marketplace.

data UsagePlan

Represents a usage plan than can specify who can assess associated API stages with specified request limits and quotas.

In a usage plan, you associate an API by specifying the API's Id and a stage name of the specified API. You add plan customers by adding API keys to the plan.

Create and Use Usage Plans

See: usagePlan smart constructor.

Response Lenses

upApiStages :: Lens' UsagePlan [APIStage]

The associated API stages of a usage plan.

upName :: Lens' UsagePlan (Maybe Text)

The name of a usage plan.

upId :: Lens' UsagePlan (Maybe Text)

The identifier of a UsagePlan resource.

upThrottle :: Lens' UsagePlan (Maybe ThrottleSettings)

The request throttle limits of a usage plan.

upQuota :: Lens' UsagePlan (Maybe QuotaSettings)

The maximum number of permitted requests per a given unit time interval.

upDescription :: Lens' UsagePlan (Maybe Text)

The description of a usage plan.

upProductCode :: Lens' UsagePlan (Maybe Text)

The AWS Markeplace product identifier to associate with the usage plan as a SaaS product on AWS Marketplace.